### 2024-06-13 #153 @editorialbot
[PRE REVIEW]: Dionysos.jl: a Modular Platform for Smart Symbolic Control
**Submitting author:** @adrienbanse (<a href="http://orcid.org/0000-0002-4456-6618">Adrien Banse</a>)
**Repository:** https://github.com/dionysos-dev/Dionysos.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
**Managing EiC:** Mathieu Besançon

Thanks for submitting your paper to JCON @adrienbanse. **Currently, there isn't a JCON editor assigned** to your paper.
