+++ problems/path_planning_sgmpc.jl
@@ -11,161 +11,153 @@ const ST = Dionysos.System
 @enum ApproxMode GROWTH LINEARIZED
 
 function dynamicofsystem()
-	# System eq x' = F_sys(x, u)
-	function F_sys(x, u)
-		return SVector{3}(
-			x[1] + u[1] * cos(x[3]),
-			x[2] + u[1] * sin(x[3]),
-			x[3] + u[2] % (2 * π),
-		)
-	end
+    # System eq x' = F_sys(x, u)
+    function F_sys(x, u)
+        return SVector{3}(
+            x[1] + u[1] * cos(x[3]),
+            x[2] + u[1] * sin(x[3]),
+            x[3] + u[2] % (2 * π),
+        )
+    end
 
-	# We define the growth bound function of $f$:
-	ngrowthbound = 5
+    # We define the growth bound function of $f$:
+    ngrowthbound = 5
 
-	# We define the growth bound function of $f$:
-	# FIXME: Double check if this is need and is the correct growth bound
-	function L_growthbound(u)
-		beta = abs(u[1]) # We assume that the velocity is bounded by 1?
-		gamma = abs(u[2])  # We assume that the angle is bounded by 2π
-		return SMatrix{3, 3}(0.0, 0.0, 0.0, 0.0, 0.0, 0.0, beta, beta, gamma)
-	end
+    # We define the growth bound function of $f$:
+    # FIXME: Double check if this is need and is the correct growth bound
+    function L_growthbound(u)
+        beta = abs(u[1]) # We assume that the velocity is bounded by 1?
+        gamma = abs(u[2])  # We assume that the angle is bounded by 2π
+        return SMatrix{3, 3}(0.0, 0.0, 0.0, 0.0, 0.0, 0.0, beta, beta, gamma)
+    end
 
-	# We define the linearized system of $f$:
-	# DF(x, u) = ∂f/∂x(x, u)
-	function DF_sys(x, u)
-		return SMatrix{3, 3}(
-			1.0,
-			0.0,
-			-u[1] * cos(x[3]),
-			0.0,
-			1.0,
-			-u[1] * sin(x[3]),
-			0.0,
-			0.0,
-			1,
-		)
-	end
+    # We define the linearized system of $f$:
+    # DF(x, u) = ∂f/∂x(x, u)
+    function DF_sys(x, u)
+        return SMatrix{3, 3}(
+            1.0,
+            0.0,
+            -u[1] * cos(x[3]),
+            0.0,
+            1.0,
+            -u[1] * sin(x[3]),
+            0.0,
+            0.0,
+            1,
+        )
+    end
 
-	# We define the bound of the linearized system of $f$:
-	bound_DF(u) = 0.0
-	bound_DDF(u) = 0.0
+    # We define the bound of the linearized system of $f$:
+    bound_DF(u) = 0.0
+    bound_DDF(u) = 0.0
 
-	return F_sys, L_growthbound, ngrowthbound, DF_sys, bound_DF, bound_DDF
+    return F_sys, L_growthbound, ngrowthbound, DF_sys, bound_DF, bound_DDF
 end
 
 function filter_obstacles(_X_, _I_, _T_, obs)
-	obstacles = typeof(_X_)[]
-	for ob in obs
-		if ob ⊆ _X_ && isempty(ob ∩ _I_) && isempty(ob ∩ _T_)
-			push!(obstacles, ob)
-		end
-	end
[…]
