+++ Project.toml
@@ -19,6 +19,7 @@ IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 IntervalLinearAlgebra = "92cbe1ac-9c24-436b-b0c9-5f7317aedcd5"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
+LatexPrint = "d2208f48-c256-5759-9089-c25ed2a93924"
 LazySets = "b4f0291d-fe17-52bc-9479-3d1a343d9043"
 Libtiff_jll = "89763e89-9b03-5906-acba-b20f662cd828"
 LightXML = "9c8b4983-aa76-5018-a973-4c85ecc9e179"
