+++ problems/pc-osqp.jl
@@ -18,19 +18,19 @@ const SY = DI.Symbolic
 
 # Define the structure for the ADP parameters:
 struct ADPParameters
-	gamma::Float64
-	N_adp::UnitRange{Int}
-	delta::Float64
-	N_tail::Int
+    gamma::Float64
+    N_adp::UnitRange{Int}
+    delta::Float64
+    N_tail::Int
 
-	function ADPParameters()
-		new(0.95, 1:5, 5.5, 50)
-	end
+    function ADPParameters()
+        return new(0.95, 1:5, 5.5, 50)
+    end
 end
 
 # Define the structure for the simulation parameters:
 struct SimulationParameters
-	Ts::Float64
+    Ts::Float64
     freq::Float64
     torque::Float64
     t0::Float64
@@ -38,317 +38,360 @@ struct SimulationParameters
     sim_periods::Int
     flag_steady_trans::Int
 
-	function SimulationParameters()
-		Ts = 25.0e-06            # Sampling time
-		freq = 50.0               # Switching frequency
-		torque = 1.0              # Desired torque
-		t0 = 0.0                 # Initial time
-		init_periods = 1         # Number of periods to settle before simulation
-		sim_periods = 2          # Numer of simulated periods
-		flag_steady_trans = 0    # Flag Steady State (0) or Transients (1)
+    function SimulationParameters()
+        Ts = 25.0e-06            # Sampling time
+        freq = 50.0               # Switching frequency
+        torque = 1.0              # Desired torque
+        t0 = 0.0                 # Initial time
+        init_periods = 1         # Number of periods to settle before simulation
+        sim_periods = 2          # Numer of simulated periods
+        flag_steady_trans = 0    # Flag Steady State (0) or Transients (1)
 
-		new(Ts, freq, torque, t0, init_periods, sim_periods, flag_steady_trans)
-	end
+        return new(Ts, freq, torque, t0, init_periods, sim_periods, flag_steady_trans)
+    end
 end
 
 # Define the structure for the parameters:
 struct Parameters
-	wr::Float64
-	is::Float64
-	Rs::Float64
-	Rr::Float64
-	Xls::Float64
-	Xlr::Float64
-	Xm::Float64
-	Vdc::Float64
-	xc::Float64
-	Vrat::Float64
-	Irat::Float64
-	RealPower::Float64
-	ApparentPower::Float64
-	RotationalSpeed::Float64
-	iVdc::Float64
-	k1::Float64
-	k2::Float64
-	fsw_des::Float64
-	kT::Float64
-	P::SMatrix
-	invP::SMatrix
-	adp_params::ADPParameters
+    wr::Float64
+    is::Float64
+    Rs::Float64
+    Rr::Float64
+    Xls::Float64
+    Xlr::Float64
+    Xm::Float64
+    Vdc::Float64
+    xc::Float64
+    Vrat::Float64
+    Irat::Float64
+    RealPower::Float64
+    ApparentPower::Float64
+    RotationalSpeed::Float64
+    iVdc::Float64
+    k1::Float64
+    k2::Float64
+    fsw_des::Float64
+    kT::Float64
+    P::SMatrix
+    invP::SMatrix
+    adp_params::ADPParameters
[…]
