### 2024-06-14 #153 @lucaferranti
Hi @jmskov :wave: ,

would you like to review this JuliaCon proceedings submission?

you can read more about reviewign for JuliaCon proceedings on the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/).
