### 2024-06-14 #160 @lucaferranti
Thank you @adrienbanse for the submission and @schillic and @jmskov for volunteering as reviewers. I will be the editor for this submission.

Reviewers, as first, you can generate your checklist with

After that, feel free to leave your review comments either here or as issues in the submission repository. You can find more information in the [review guide](https://juliacon.github.io/proceedings-guide/reviewer/) and feel free to ping me at any time if you have questions.
