+++ paper/ref.bib
@@ -8,35 +8,40 @@
   number={Special Centennial Issue},
   pages={1287--1308},
   year={2012},
-  publisher={IEEE}
+  publisher={IEEE}, 
+  doi = {10.1109/JPROC.2012.2189792}
 }
 
 @book{alur2015principles,
   title={Principles of cyber-physical systems},
   author={Alur, Rajeev},
   year={2015},
-  publisher={MIT press}
+  publisher={MIT press}, 
+  doi = {10.1017/9781107588981}
 }
 
 @book{lee2016introduction,
   title={Introduction to embedded systems: A cyber-physical systems approach},
   author={Lee, Edward Ashford and Seshia, Sanjit Arunkumar},
   year={2016},
-  publisher={MIT press}
+  publisher={MIT press}, 
+  doi={10.1145/1719010.1719011}
 }
 
 @book{baier2008principles,
   title={Principles of model checking},
   author={Baier, Christel and Katoen, Joost-Pieter},
   year={2008},
-  publisher={MIT press}
+  publisher={MIT press},
+  doi={10.5555/1373322}
 }
 
 @book{tabuada2009verification,
   title={Verification and control of hybrid systems: a symbolic approach},
   author={Tabuada, Paulo},
   year={2009},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media}, 
+  doi={10.1007/978-1-4419-0224-5}
 }
 
 @inproceedings{alur1998alternating,
@@ -45,7 +50,8 @@
   booktitle={CONCUR'98 Concurrency Theory: 9th International Conference Nice, France, September 8--11, 1998 Proceedings 9},
   pages={163--178},
   year={1998},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/BFb0055622}
 }
 
 @article{reissig2016feedback,
@@ -56,7 +62,8 @@
   number={4},
   pages={1781--1796},
   year={2016},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/CDC.2014.7039364}
 }
 
 @book{belta2017formal,
@@ -64,7 +71,8 @@
   author={Belta, Calin and Yordanov, Boyan and Gol, Ebru Aydin},
   volume={89},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-319-50763-7}
 }
 
 @article{kupferman2001model,
@@ -74,7 +82,8 @@
   volume={19},
   pages={291--314},
   year={2001},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1023/A:1011254632723}
 }
 
 @article{angeli2002lyapunov,
@@ -85,7 +94,8 @@
   number={3},
   pages={410--421},
   year={2002},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/9.989067}
 }
 
 @article{lohmiller1998contraction,
@@ -96,7 +106,8 @@
   number={6},
   pages={683--696},
   year={1998},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/S0005-1098(98)00019-3},
 }
 
 @inproceedings{camara2011safety,
@@ -105,7 +116,8 @@
   booktitle={2011 50th IEEE Conference on Decision and Control and European Control Conference},
   pages={520--525},
   year={2011},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/CDC.2011.6160424}
 }
 
 @inproceedings{hsu2018multi,
@@ -113,7 +125,8 @@
[…]
