+++ paper/ref.bib
@@ -32,8 +32,7 @@
   title={Principles of model checking},
   author={Baier, Christel and Katoen, Joost-Pieter},
   year={2008},
-  publisher={MIT press},
-  doi={10.5555/1373322}
+  publisher={MIT press}
 }
 
 @book{tabuada2009verification,
@@ -107,7 +106,7 @@
   pages={683--696},
   year={1998},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/S0005-1098(98)00019-3},
+  doi = {10.1016/S0005-1098(98)00019-3},
 }
 
 @inproceedings{camara2011safety,
