### 2024-06-14 #160 @schillic
@adrienbanse: The problem is the `{\_}` in the doi entries. Try `_` or, if that does not work, `\,`.
