### 2024-06-15 #160 @lucaferranti
@adrienbanse RE: DOI

- For the missing ones, if the references don't have a DOI, then it's ok

- The invalid ones should be fixed, that is generally due to some latex formatting quirkiness and it generally has to do with some underscores, so I think Christian's suggestion is on the money
