+++ paper/ref.bib
@@ -147,7 +147,7 @@
   pages={351--365},
   year={2009},
   organization={Springer},
-  doi={10.1007/978-3-642-00602-9{\_}25}
+  doi={10.1007/978-3-642-00602-9\_25}
 }
 
 %%%%%%%%%%%%%%%%%% Toolboxes
@@ -158,7 +158,7 @@
   pages={566--569},
   year={2010},
   organization={Springer},
-  doi={10.1007/978-3-642-14295-6{\_}49}
+  doi={10.1007/978-3-642-14295-6\_49}
 }
 
 @inproceedings{rungger2016scots,
