### 2024-06-17 d63b0b3
paper: fix doi, \textunderscore
