### 2024-06-17 #160 @adrienbanse
@blegat In our paper, the `underscore` package breaks all the equations using subscripts, also just tried with `\textunderscore` and it's not allowed by the editorial manager either
I'll check for similar issues in previous reviews
