+++ paper/paper.tex
@@ -3,6 +3,7 @@
 \setcounter{page}{1}
 
 % added
+\usepackage{underscore}
 \usepackage{amsmath}
 
 \begin{document}
