### 2024-06-17 #160 @schillic
@adrienbanse I just checked two old papers of mine and there `\_` used to work. I do not know what changed here.
