### 2024-06-17 #160 @lucaferranti
maybe something changed when upgrading from `whedon` to `editorialbot`? I can check on how our references process differs from JOSS and see if I spot something
