### 2024-06-17 #160 @adrienbanse
@schillic @lucaferranti The last Whedon code I found (but archived) just appends the string written in the bib file to `https://doi.org/` and queries the url, hence the error if it's anything but an underscore

See https://github.com/openjournals/whedon-api/blob/ad17771bc842c1e29fdddc4978c242dbc979454f/workers.rb#L289
```
  def invalid_doi?(doi_string)
    return true if doi_string.nil?

    if doi_string.include?('http')
      return true
    else
      url = "https://doi.org/#{doi_string}"
    end

    escaped_url = URI.escape(url)

    begin
      status_code = Faraday.head(escaped_url).status
      if [301, 302].include? status_code
        return false
      else
        return true
      end
    rescue Faraday::ConnectionFailed
      return true
    rescue URI::InvalidURIError
      return true
    end
  end
```
