+++ .github/workflows/documentation.yml
@@ -16,15 +16,15 @@ jobs:
       with:
         # Build documentation on the latest Julia 1.x
         version: '1'
-      - name: Install dependencies
-        shell: julia --project=docs/ {0}
-        run: |
-          using Pkg
-          Pkg.add([
-              PackageSpec(url="https://github.com/blegat/IntervalLinearAlgebra.jl", rev="radius"),
-              PackageSpec(path=pwd()),
-          ])
-          Pkg.instantiate()
+    - name: Install dependencies
+      shell: julia --project=docs/ {0}
+      run: |
+        using Pkg
+        Pkg.add([
+            PackageSpec(url="https://github.com/blegat/IntervalLinearAlgebra.jl", rev="radius"),
+            PackageSpec(path=pwd()),
+        ])
+        Pkg.instantiate()
     - name: Build and deploy
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # For authentication with GitHub Actions token
