### 2024-06-19 709c818
fix: init set lower and upper bounds
