+++ problems/path_planning_sgmpc.jl
@@ -174,13 +174,13 @@ function problem(; simple = false, approx_mode::ApproxMode = GROWTH)
     if simple
         _X_ = UT.HyperRectangle(SVector(-3.5, -2.6, -pi), SVector(3.5, 2.6, pi))
         _I_ = UT.HyperRectangle(SVector(1.0, -1.7, 0.0), SVector(1.2, -1.5, 0.0))
-        _T_ = UT.HyperRectangle(SVector(0.5, 0.5, 0.0), SVector(0.7, 0.7, 0.0))
+        _T_ = UT.HyperRectangle(SVector(0.5, 0.5, -pi), SVector(0.7, 0.7, pi))
     else
         _X_ = UT.HyperRectangle(SVector(-3.5, -2.6, -pi), SVector(3.5, 2.6, pi))
         _I_ = UT.HyperRectangle(SVector(1.0, -1.7, 0.0), SVector(1.2, -1.5, 0.0))
         _T_ = UT.HyperRectangle(
-            SVector(sqrt(32.0 / 3.0), sqrt(20.0 / 3.0), 0.0),
-            SVector(sqrt(32.0 / 3.0) - 0.2, sqrt(20.0 / 3.0) - 0.2, 0.0),
+            SVector(sqrt(32.0 / 3.0), sqrt(20.0 / 3.0), -pi),
+            SVector(sqrt(32.0 / 3.0) - 0.2, sqrt(20.0 / 3.0) - 0.2, pi),
         )
     end
     obs = get_obstacles(_X_)
