### 2024-06-19 85b3969
fix: control origin and initial state
