### 2024-06-27 d56026d
implementing the resetmaps
