+++ problems/pc-osqp-rmaps.jl
@@ -1,0 +1,177 @@
+module PCOSQPRM
+using Test
+
+using Dionysos: Dionysos
+const DI = Dionysos
+const UT = DI.Utils
+const PR = DI.Problem
+
+using FillArrays
+using Polyhedra
+using MathematicalSystems, HybridSystems
+using SemialgebraicSets
+using LinearAlgebra
+
+using CDDLib: CDDLib
+
+function system(lib, T::Type)
+
+    # Generate the automaton nodes
+    possible_values = [[-1, 0, 1], [-1, 0, 1], [-1, 0, 1], [0, 1], [0, 1], [0, 1]]
+    nvars = length(possible_values)
+
+    ## generate all possible combinations of the values
+    function generate_combinations(possible_values, depth)
+        if depth == 1
+            return [[v] for v in possible_values[1]]
+        end
+        return [
+            [v; c...] for v in possible_values[1] for c in generate_combinations(possible_values[2:end], depth - 1)
+        ]
+    end
+
+    ## generate all possible combinations of the values
+    nodes = generate_combinations(possible_values, nvars)
+    nnodes = length(nodes)
+    
+	domains = []
+
+	#Automaton:
+	automaton = GraphAutomaton(nnodes)
+    indexB = Int[]
+
+	#Resetmaps: (guards + reset)
+	A = T[
+		0.9999981424449595 9.91220860939534e-12 3.4481569821196496e-7 9.205526977551878e-5 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
+		-9.91220860939534e-12 0.9999981424449595 -9.205526977551878e-5 3.4481569821196496e-7 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
+		2.1535327064414276e-7 -2.644213073804883e-12 0.9999999080158245 -2.455696808386769e-5 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
+		2.644213073804883e-12 2.1535327064414276e-7 2.455696808386769e-5 0.9999999080158245 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
+		0.0 0.0 0.0 0.0 0.9999999996930378 -2.477749999746475e-5 0.0 0.0 0.0 0.0 0.0 0.0
+		0.0 0.0 0.0 0.0 2.477749999746475e-5 0.9999999996930378 0.0 0.0 0.0 0.0 0.0 0.0
+		0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
+		0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
+		0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0
+		0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.99875 0.0 0.0
+		0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0012499999999999734 0.99875 0.0
+		0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 1.0
+	]
+
+	B = T[
+		6.256252307659102e-5 -3.128126153811649e-5 -3.128126153847453e-5 0.0 0.0 0.0
+		-2.0671108273679513e-16 5.418073430928138e-5 -5.418073430907468e-5 0.0 0.0 0.0
[…]
