### 2024-07-05 9ee82a3
adding the optimal function
