+++ previews/PR354/generated/Path planning sgmpc.jl
@@ -14,12 +14,12 @@ include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "path_planning_
 
 initial = SVector(1.0, -1.7, 0.0)
 #initial = SVector(0.0, 0.0, 0.0)
-target = SVector(0.5, 0.5, -pi)
+#target = SVector(0.5, 0.5, -pi)
 #target = SVector(2.6, 2.0, -pi)
 #target = SVector(-2.6, 2.0, -pi)
-#target = SVector(sqrt(32.0 / 3.0), sqrt(20.0 / 3.0), -pi)
+target = SVector(sqrt(32.0 / 3.0), sqrt(20.0 / 3.0), -pi)
 
-concrete_problem = PathPlanningSgMPC.problem(; sgmpc = false, initial = initial, target = target)
+concrete_problem = PathPlanningSgMPC.problem(; sgmpc = true, initial = initial, target = target)
 concrete_system = concrete_problem.system;
 
 x0 = SVector(0.0, 0.0, 0.0);
