+++ previews/PR354/examples/solvers/Path planning sgmpc.jl
@@ -63,13 +63,13 @@ include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "path_planning_
 # Defining the initial and target sets for the state-space of the system:
 initial = SVector(1.0, -1.7, 0.0)
 #initial = SVector(0.0, 0.0, 0.0)
-target = SVector(0.5, 0.5, -pi)
+#target = SVector(0.5, 0.5, -pi)
 #target = SVector(2.6, 2.0, -pi)
 #target = SVector(-2.6, 2.0, -pi)
-#target = SVector(sqrt(32.0 / 3.0), sqrt(20.0 / 3.0), -pi)
+target = SVector(sqrt(32.0 / 3.0), sqrt(20.0 / 3.0), -pi)
 
 # Now we instantiate the problem using the function provided by [PathPlanning.jl](@__REPO_ROOT_URL__/problems/PathPlanningSgMPC.jl) 
-concrete_problem = PathPlanningSgMPC.problem(; sgmpc = false, initial = initial, target = target)
+concrete_problem = PathPlanningSgMPC.problem(; sgmpc = true, initial = initial, target = target)
 concrete_system = concrete_problem.system;
 
 # ### Definition of the abstraction
