+++ problems/path_planning_sgmpc.jl
@@ -8,40 +8,38 @@ const DO = Dionysos.Domain
 const PB = Dionysos.Problem
 const ST = Dionysos.System
 
-
 # TODO: Move to Dionysos.Utils (the function is already there, but it imposes that Q, R, and N must be the same type and *size* and that q and r should be the same type and *size*)
-struct QuadraticStateControlFunction{T} <:
-    UT.ScalarControlFunction
- Q::AbstractMatrix{T}
- R::AbstractMatrix{T}
- N::AbstractMatrix{T}
- q::AbstractArray{T}
- r::AbstractArray{T}
- v::T
+struct QuadraticStateControlFunction{T} <: UT.ScalarControlFunction
+    Q::AbstractMatrix{T}
+    R::AbstractMatrix{T}
+    N::AbstractMatrix{T}
+    q::AbstractArray{T}
+    r::AbstractArray{T}
+    v::T
 
- #=
- function QuadraticStateControlFunction(Q::MT, R::MT, N::MT, q::AT, r::AT, v::T) where {T, MT <: AbstractMatrix{T}, AT <: AbstractArray{T}}
-    # Perform size checks
-    @assert size(Q, 1) == size(Q, 2) "Q must be square"
-    @assert size(R, 1) == size(R, 2) "R must be square"
-    @assert size(Q, 1) == size(N, 1) "Q and N must have compatible dimensions"
-    @assert size(R, 1) == size(N, 2) "R and N must have compatible dimensions"
-    @assert size(Q, 1) == length(q) "Q and q must have compatible dimensions"
-    @assert size(R, 1) == length(r) "R and r must have compatible dimensions"
-    new{T, MT, AT}(Q, R, N, q, r, v)
-end
-=#
+    #=
+    function QuadraticStateControlFunction(Q::MT, R::MT, N::MT, q::AT, r::AT, v::T) where {T, MT <: AbstractMatrix{T}, AT <: AbstractArray{T}}
+       # Perform size checks
+       @assert size(Q, 1) == size(Q, 2) "Q must be square"
+       @assert size(R, 1) == size(R, 2) "R must be square"
+       @assert size(Q, 1) == size(N, 1) "Q and N must have compatible dimensions"
+       @assert size(R, 1) == size(N, 2) "R and N must have compatible dimensions"
+       @assert size(Q, 1) == length(q) "Q and q must have compatible dimensions"
+       @assert size(R, 1) == length(r) "R and r must have compatible dimensions"
+       new{T, MT, AT}(Q, R, N, q, r, v)
+    end
+    =#
 
 end
 function function_value(f::QuadraticStateControlFunction, x, u)
- return x'f.Q * x + u'f.R * u + 2 * (x'f.N * u + x'f.q + u'f.r) + f.v
+    return x'f.Q * x + u'f.R * u + 2 * (x'f.N * u + x'f.q + u'f.r) + f.v
 end
 function get_full_psd_matrix(f::QuadraticStateControlFunction)
- return [
-     f.Q f.N f.q
-     f.N' f.R f.r
-     f.q' f.r' f.v
- ]
+    return [
+        f.Q f.N f.q
+        f.N' f.R f.r
+        f.q' f.r' f.v
+    ]
 end
 ##########
 
@@ -115,9 +113,9 @@ function system(
     udim = 2,
     sysnoise = SVector(0.0, 0.0, 0.0),
     measnoise = SVector(0.0, 0.0, 0.0),
-    tstep = 1.,
+    tstep = 1.0,
     nsys = 5,
-    ngrowthbound = 5
+    ngrowthbound = 5,
 )
     sys_map = let nsys = nsys
[…]
