### 2024-07-06 #160 @jmskov
The package itself looks very useful, and the examples in the paper are compelling. I have some ideas for incorporating probabilistic and unbounded disturbance sets, but that is for another time.

- Currently, a couple of tests fail for me on MacOS on fresh Julia 1.6 and 1.10 in different ways. On 1.6, lines 213 and 214 of `test_ellipsoids_abstraction.jl` fail the approximation check, with `Expression: ≈(cost_bound, 0.6250139513432214, rtol = 0.001)`. A MacOS+RNG issue? On 1.10, there is a conflict with `radius` with LazySets.jl. Perhaps a version issue.
- Running `benchmark.jl`, I get an undefined `SCOTSAbstraction`, which I can't seem to find in the relevant module.
- (I will update the above comments if I am able to resolve)
- No explicit "Statement of need" section in the paper
- For the problem definition, `time::T`, is the number of allowed steps. Can it handle a duration input that determines the maximum number of discrete steps?
- Perhaps outside the scope of this package, but adding the capability to return a cost or another measure (probability eh?) with the controller would be useful for the future
