### 2024-07-06 #160 @blegat
Thanks for the review! About the issue with `radius`, it is fixed by https://github.com/JuliaIntervals/IntervalLinearAlgebra.jl/pull/135. At the moment, you need to checkout this branch of `IntervalLinearAlgebra`
