### 2024-07-06 #160 @schillic
I did not yet try out the library itself (will do that in the near future). I have filled out the checklist (above) and below I have some comments on the paper. There is only one major comment, which I hope you can fix or explain.
Overall a really nice paper (and package - from what I can tell so far) :+1:

---

#### Summary

The paper introduces Dionysos.jl, a Julia library for controller synthesis based on abstraction. The library has a modular structure, offering systems, problems, algorithms, set representations, etc. Currently, a key feature of Dionysos are methods based on lazy and non-partition abstraction of the state space. The usage is explained with an extensive example, and the capabilities are demonstrated in a comparison to SCOTS and CoSyMA (which is available on CodeOcean - which I think is a great effort!).

#### Main comment

- There is a mismatch between Definition 3, defined over infinite words, and the behavior $B(S)$, which can be finite or infinite. Since specifications require only infinite behavior, but every infinite behavior induces finite prefixes, that makes any behavior violate any specification.

#### Checklist comments

- **Installation instructions**: The README points to https://julialang.github.io/Pkg.jl/v1/managing-packages/#Adding-unregistered-packages, but the package is registered. This should be updated, probably to https://dionysos-dev.github.io/Dionysos.jl/stable/developers/setup/.
- **A statement of need**: While there is no dedicated section, the purpose of the package is clearly communicated.

#### Minor comments

- p. 1: "solving optimal control problems using state-of-the-art control techniques, from control" - I guess the 2nd "control" could be dropped.
- p. 2: "integer numbers" -> just "integers"?
- p. 2 last paragraph on the left: Should it be $k$ ∈ { $0, ..., T-1$ }?
- It may be better to move Fig. 2 to the next page so that it is closer to the corresponding text.
- Fig. 2 and p. 3: I do not see why a strict single-valued map must induce a partition. Could $R_1$ not map all states $a$ to the same state $b$?
- Fig. 3: I do not understand where the red box comes from. Consequently, it is unclear to me why the two cells on the right are also red. Is that an overestimate of $F_1$?
- Fig. 4: What does $\kappa \in U_2$ mean? $\kappa$ is a function, right?
- p. 4: "specified specifications" - a bit convoluted
- p. 4: "cost. :"
- p. 4: ", it complements" -> "and complements"
- Section 4.2 mentions that obstacles are encoded as part of the domain. Maybe it is worth mentioning again that the Domain module is not described in the paper.
- p. 5: SafetyProblem: "the set $X_I$" -> "the initial set $X_I$" (for consistency)
- p. 6: Consider merging the definition of an ellipsoid (8) with the one at the end of Section 3.
- p. 6: "LazyEllispoidsAbstraction" -> "LazyEllipsoidsAbstraction"
- Could you expand the explanation of Fig. 7? What does the arrow from the white center of the figure mean? And do I understand correctly that no further abstract states are reachable? Maybe you can also highlight the initial and target sets (in a different color), like in the web tutorial.
- p. 7: "Comparaison" -> "Comparison"
