### 2024-07-09 #160 @schillic
I had a look at the code and have some comments:

#### Already reported, but for completeness

- I confirm the failing test (at least one, did not check for more).
- `WARNING: both LazySets and IntervalArithmetic export "radius"; uses of it in module IntervalLinearAlgebra must be qualified`
   - This is really a problem because I cannot run the example in the paper without a crash:

    ERROR: UndefVarError: `radius` not defined

#### Further warnings

- `WARNING: using LazySets.vertices in module Utils conflicts with an existing identifier.`
- There are some duplicate method definitions (presumably from a merge conflict) in `src/optim/abstraction/`:
```
WARNING: Method definition (::Type{Dionysos.Optim.Abstraction.LazyAbstraction.Optimizer{T} where T})() in module LazyAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_abstraction.jl:66 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_abstraction.jl:87.
WARNING: Method definition is_empty(Dionysos.Optim.Abstraction.LazyAbstraction.Optimizer{T} where T) in module LazyAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_abstraction.jl:68 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_abstraction.jl:89.
WARNING: Method definition set(Dionysos.Optim.Abstraction.LazyAbstraction.Optimizer{T} where T, MathOptInterface.RawOptimizerAttribute, Any) in module LazyAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_abstraction.jl:70 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_abstraction.jl:91.
WARNING: Method definition get(Dionysos.Optim.Abstraction.LazyAbstraction.Optimizer{T} where T, MathOptInterface.RawOptimizerAttribute) in module LazyAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_abstraction.jl:83 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_abstraction.jl:94.
```
```
WARNING: Method definition (::Type{Dionysos.Optim.Abstraction.HierarchicalAbstraction.Optimizer{T} where T})() in module HierarchicalAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/hierarchical_abstraction.jl:75 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/hierarchical_abstraction.jl:88.
WARNING: Method definition is_empty(Dionysos.Optim.Abstraction.HierarchicalAbstraction.Optimizer{T} where T) in module HierarchicalAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/hierarchical_abstraction.jl:77 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/hierarchical_abstraction.jl:90.
WARNING: Method definition set(Dionysos.Optim.Abstraction.HierarchicalAbstraction.Optimizer{T} where T, MathOptInterface.RawOptimizerAttribute, Any) in module HierarchicalAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/hierarchical_abstraction.jl:79 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/hierarchical_abstraction.jl:92.
```
```
WARNING: Method definition (::Type{Dionysos.Optim.Abstraction.LazyEllipsoidsAbstraction.Optimizer{T} where T})() in module LazyEllipsoidsAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_ellipsoids_abstraction.jl:84 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_ellipsoids_abstraction.jl:101.
WARNING: Method definition is_empty(Dionysos.Optim.Abstraction.LazyEllipsoidsAbstraction.Optimizer{T} where T) in module LazyEllipsoidsAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_ellipsoids_abstraction.jl:86 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_ellipsoids_abstraction.jl:103.
WARNING: Method definition set(Dionysos.Optim.Abstraction.LazyEllipsoidsAbstraction.Optimizer{T} where T, MathOptInterface.RawOptimizerAttribute, Any) in module LazyEllipsoidsAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_ellipsoids_abstraction.jl:88 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_ellipsoids_abstraction.jl:105.
WARNING: Method definition get(Dionysos.Optim.Abstraction.LazyEllipsoidsAbstraction.Optimizer{T} where T, MathOptInterface.RawOptimizerAttribute) in module LazyEllipsoidsAbstraction at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_ellipsoids_abstraction.jl:97 overwritten at ~/.julia/dev/Dionysos/src/optim/abstraction/lazy_ellipsoids_abstraction.jl:111.
```
