### 2024-08-12 6527a25
CompatHelper: bump compat for Symbolics to 6, (keep existing compat)
