+++ Project.toml
@@ -69,7 +69,7 @@ SimpleWeightedGraphs = "1.2"
 SpecialFunctions = "2"
 StaticArrays = "1"
 Suppressor = "0.2"
-Symbolics = "5"
+Symbolics = "5, 6"
 julia = "1.6"
 
 [extras]
