+++ paper/paper.tex
@@ -3,7 +3,7 @@
 \setcounter{page}{1}
 
 % added
-\usepackage{underscore}
+%\usepackage{underscore}
 \usepackage{amsmath}
 
 \begin{document}
@@ -41,7 +41,7 @@ Several open-source toolboxes~\cite{mazo2010pessoa,Roy2011,rungger2016scots} hav
 
 \vskip 6pt
 %%% The (software) solution
-To address this issue, we introduce \texttt{Dionysos.jl}\footnote{The package is open source and available on GitHub at: \url{https://github.com/dionysos-dev/Dionysos.jl}.}, a modular Julia \cite{bezanson2017julia} package for solving optimal control problems using state-of-the-art control techniques, from control, optimization, and machine learning, for complex systems.
+To address this issue, we introduce \texttt{Dionysos.jl}\footnote{The package is open source and available on GitHub at: \url{https://github.com/dionysos-dev/Dionysos.jl}.}, a modular Julia \cite{bezanson2017julia} package for solving optimal control problems using state-of-the-art techniques, from control, optimization, and machine learning, for complex systems.
 It is built on top of different Julia packages such as \texttt{JuMP.jl} and \texttt{MathOptInterface.jl}, and
 features optimal control problem definitions and several abstraction-based methods to solve them.
 %% Origin of the project: ERC
@@ -75,14 +75,16 @@ This paper is structured as follows.  Section~\ref{sec:abstraction} introduces t
 \vskip 6pt
 \noindent \textbf{Notation:} 
 % Intervals
-The sets $\R,\Z, \Z_+$ denote respectively the sets of real numbers, integers and non-negative integer numbers.
+The sets $\R,\Z, \Z_+$ denote respectively the sets of real numbers, integers and non-negative integers.
+% Kleene closure
+\color{red}Given a set $A$, the set $A^*$ is its \emph{Kleene closure}. \color{black}
 % Set-valued map
 Given two sets $A,B$, we define a \emph{single-valued map} as $f:A\rightarrow B$, while a~\emph{set-valued map} is defined as $f:A\rightarrow 2^B$, where $2^B$ is the power set of $B$, i.e., the set of all subsets of $B$. 
 The image of a subset $\Omega\subseteq A$ under $f:A\rightarrow 2^B$ is denoted $f(\Omega)$.
 % Relation
 We identify a binary \emph{relation} $R\subseteq A \times B$ with set-valued maps, i.e., $R(a) = \{b \mid (a, b)\in R\}$ and $R^{-1}(b) =\{ a \mid (a, b)\in R\}$.
 %
-A relation $R\subseteq A\times B$ is \emph{strict} (resp. \emph{single-valued)} if for every $a\in A$ the set $R(a)\ne \myemptyset$ (resp. $R(a)$ is a singleton). 
+A relation $R\subseteq A\times B$ is \emph{strict} (resp. \emph{single-valued}) if for every $a\in A$ the set $R(a)\ne \myemptyset$ (resp. $R(a)$ is a singleton). 
 
 \section{Abstraction-based control} \label{sec:abstraction}
 
@@ -121,7 +123,7 @@ A \emph{finite-state} system, in contrast to an \emph{infinite-state} system, re
 
 \vskip 6pt
 % Definition of a solution
[…]
