### 2024-08-28 #160 @lucaferranti
Hi everyone :wave:  ,

just checking in on how this is going.

@blegat have you addressed the comments from the reviewer (the issue in IntervalLinearAlgebra.jl should be fixed now)? If yes, feel free to ping them to take a second look? Otherwise, do you have an estimate timeline?
