### 2024-08-28 #160 @blegat
Thanks for the help with IntervalLinearAlgebra. We're still working on addressing the reviews, see https://github.com/dionysos-dev/Dionysos.jl/pull/361.
