+++ paper/paper.tex
@@ -5,7 +5,7 @@
 % added
 %\usepackage{underscore}
 \usepackage{amsmath}
-
+\usepackage{subcaption}
 \begin{document}
 
 \input{header}
@@ -173,9 +173,11 @@ Notably, the condition that $R$ is a strict relation is equivalent to ensuring t
 
 \begin{figure}[ht!]
 \centerline{\includegraphics[width=8.5cm]{Figures/Abstraction/strict-cover-partition.pdf}}
-\caption{Types of discretization of the concrete state space.
+\caption{
+    Types of discretization of the concrete state space.
     Let $\Sys_{1} = (\set{X}_1,\set{U}_1,F_1)$ with $\set{X}_1 = [0,1]^2$, $\Sys_{2} = (\set{X}_2,\set{U}_2,F_2)$ with $\set{X}_2 = \{q_1,q_2,q_3,q_4\}$, $R_1 \subseteq \set{X}_1\times \set{X}_2$ and $R_2 \subseteq \set{X}_1\times \set{X}_2$ are explicit from the figure.
-    Left: $R_1$ is a strict single-valued map, i.e., it induces a full partition of $\set{X}_1$. Right: $R_2$ is a non-strict set-valued map, i.e., it induces a partial cover of $\set{X}_1$.}
+    Left: $R_1$ is a strict single-valued map, i.e., it induces a full partition of $\set{X}_1$. Right: $R_2$ is a non-strict set-valued map, i.e., it induces a partial cover of $\set{X}_1$.
+    }
     \label{fig:discretization}
 \end{figure}
 
@@ -203,7 +205,7 @@ Observe that the requirements $(i)$ and $(ii)$ in Definition~\ref{def:FRR} restr
 
 The classical abstraction-based approach consists in constructing a feedback refinement relation $R$ based on a predefined partition of the entire state space. 
 The limitation to piecewise constant controllers, combined with the use of a rigid/predefined partition of the entire concrete state space, could result in an intractable, or even unsolvable, abstract problem $(\Sys_2, \Sigma_2)$.
-Indeed, if the system does not exhibit local incremental stability~\cite[Definition 2.1]{angeli2002lyapunov}\cite{lohmiller1998contraction} around a cell, meaning that trajectories move away from each other, the use of piecewise constant controllers introduces a significant amount of non-determinism into the abstraction as illustrated in Figure~\ref{fig:Piece-wise-constant}. 
+Indeed, if the system does not exhibit local incremental stability~\cite[Definition 2.1]{angeli2002lyapunov}\cite{lohmiller1998contraction} around a cell, meaning that trajectories move away from each other, the use of piecewise constant controllers introduces a significant amount of non-determinism into the abstraction as illustrated in Figure~\ref{fig:Piece-wise-input}. 
 %
 Indeed, this results in a high cardinality of the set of 
 outputs of the transition map $F_2(x_2,u)$ of the abstraction.
@@ -231,21 +233,17 @@ This relation provides a simple concretization scheme, even in the presence of o
 $$\Cont_1(x_1) = (\Cont_2\circ_{I_R} R)(x_1) =\bigcup_{x_2\in R(x_1)} I_R(x_1,x_2,\Cont_2(x_2)).$$
 
 
-By designing these local state-dependent controllers, for example by solving an optimization problem (e.g.,~\cite[Section V]{calbert2024smart}), we can ensure deterministic transitions in the abstraction, thereby eliminating the non-determinism imposed by the discretization of the concrete system (see Figure~\ref{fig:Piece-wise-dependent}). 
[…]
