+++ src/system/system.jl
@@ -3,6 +3,7 @@ module System
 using StaticArrays
 using Symbolics
 using MathematicalSystems
+using LinearAlgebra
 using IntervalArithmetic
 using IntervalLinearAlgebra
 using Plots
