### 2024-09-02 2347ed7
fix: LazySets.vertices conflict warning
