+++ src/optim/abstraction/hierarchical_abstraction.jl
@@ -72,10 +72,6 @@ mutable struct Optimizer{T} <: MOI.AbstractOptimizer
     end
 end
 
-Optimizer() = Optimizer{Float64}()
-
-MOI.is_empty(optimizer::Optimizer) = optimizer.concrete_problem === nothing
-
 function MOI.set(model::Optimizer, param::MOI.RawOptimizerAttribute, value)
     if param.name == "concrete_problem"
         if !(value isa PR.OptimalControlProblem)
@@ -89,9 +85,6 @@ Optimizer() = Optimizer{Float64}()
 
 MOI.is_empty(optimizer::Optimizer) = optimizer.concrete_problem === nothing
 
-function MOI.set(model::Optimizer, param::MOI.RawOptimizerAttribute, value)
-    return setproperty!(model, Symbol(param.name), value)
-end
 function MOI.get(model::Optimizer, param::MOI.RawOptimizerAttribute)
     return getproperty(model, Symbol(param.name))
 end
