### 2024-09-09 febf2d1
Paper reviews implementation (#361)
* schillic paper review implementation

* implementation reviews julien

* test: fix tentative

* revert

* fix: LazySets.vertices conflict warning

* fix: remove duplicates due to merge conflict

* bump to julia 1.10

* CI: LTS becomes 1.10

---------

Co-authored-by: JulienCalbert <julien.calbert@uclouvain.be>
