+++ Project.toml
@@ -70,7 +70,7 @@ SpecialFunctions = "2"
 StaticArrays = "1"
 Suppressor = "0.2"
 Symbolics = "5"
-julia = "1.6"
+julia = "1.10"
 
 [extras]
 CDDLib = "3391f64e-dcde-5f30-b752-e11513730f60"
