+++ paper/paper.tex
@@ -3,7 +3,7 @@
 \setcounter{page}{1}
 
 % added
-%\usepackage{underscore}
+\usepackage{underscore}
 \usepackage{amsmath}
 \usepackage{subcaption}
 \begin{document}
