+++ paper/ref_u.bib
@@ -1,458 +1,0 @@
-
-%%%%%%%%%%%%%%%%%% Theoretical references
-@article{kim2012cyber,
-  title={Cyber--physical systems: A perspective at the centennial},
-  author={Kim, Kyoung-Dae and Kumar, Panganamala R},
-  journal={Proceedings of the IEEE},
-  volume={100},
-  number={Special Centennial Issue},
-  pages={1287--1308},
-  year={2012},
-  publisher={IEEE}, 
-  doi = {10.1109/JPROC.2012.2189792}
-}
-
-@book{alur2015principles,
-  title={Principles of cyber-physical systems},
-  author={Alur, Rajeev},
-  year={2015},
-  publisher={MIT press}, 
-  doi = {10.1017/9781107588981}
-}
-
-@book{lee2016introduction,
-  title={Introduction to embedded systems: A cyber-physical systems approach},
-  author={Lee, Edward Ashford and Seshia, Sanjit Arunkumar},
-  year={2016},
-  publisher={MIT press}, 
-  doi={10.1145/1719010.1719011}
-}
-
-@book{baier2008principles,
-  title={Principles of model checking},
-  author={Baier, Christel and Katoen, Joost-Pieter},
-  year={2008},
-  publisher={MIT press}
-}
-
-@book{tabuada2009verification,
-  title={Verification and control of hybrid systems: a symbolic approach},
-  author={Tabuada, Paulo},
-  year={2009},
-  publisher={Springer Science \& Business Media}, 
-  doi={10.1007/978-1-4419-0224-5}
-}
-
-@inproceedings{alur1998alternating,
-  title={Alternating refinement relations},
-  author={Alur, Rajeev and Henzinger, Thomas A and Kupferman, Orna and Vardi, Moshe Y},
-  booktitle={CONCUR'98 Concurrency Theory: 9th International Conference Nice, France, September 8--11, 1998 Proceedings 9},
-  pages={163--178},
-  year={1998},
-  organization={Springer},
-  doi={10.1007/BFb0055622}
-}
-
-@article{reissig2016feedback,
-  title={Feedback refinement relations for the synthesis of symbolic controllers},
-  author={Reissig, Gunther and Weber, Alexander and Rungger, Matthias},
-  journal={IEEE Transactions on Automatic Control},
-  volume={62},
-  number={4},
-  pages={1781--1796},
-  year={2016},
-  publisher={IEEE},
-  doi={10.1109/CDC.2014.7039364}
-}
-
-@book{belta2017formal,
-  title={Formal methods for discrete-time dynamical systems},
-  author={Belta, Calin and Yordanov, Boyan and Gol, Ebru Aydin},
-  volume={89},
-  year={2017},
-  publisher={Springer},
-  doi={10.1007/978-3-319-50763-7}
-}
-
-@article{kupferman2001model,
-  title={Model checking of safety properties},
-  author={Kupferman, Orna and Vardi, Moshe Y},
-  journal={Formal methods in system design},
-  volume={19},
-  pages={291--314},
-  year={2001},
-  publisher={Springer},
-  doi={10.1023/A:1011254632723}
-}
-
-@article{angeli2002lyapunov,
-  title={A {Lyapunov} approach to incremental stability properties},
-  author={Angeli, David},
-  journal={IEEE Transactions on Automatic Control},
-  volume={47},
-  number={3},
-  pages={410--421},
-  year={2002},
-  publisher={IEEE},
-  doi={10.1109/9.989067}
-}
-
-@article{lohmiller1998contraction,
-  title={On contraction analysis for non-linear systems},
-  author={Lohmiller, Winfried and Slotine, Jean-Jacques E},
-  journal={Automatica},
-  volume={34},
[…]
