+++ paper/paper.tex
@@ -5,7 +5,6 @@
 % added
 \usepackage{underscore}
 \usepackage{amsmath}
-\usepackage{subcaption}
 \begin{document}
 
 \input{header}
