### 2024-09-10 16ec35b
CompatHelper: bump compat for Symbolics to 6, (keep existing compat) (#363)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
