### 2024-09-10 #160 @adrienbanse
Many thanks @jmskov and @schillic for the time you spent reviewing this paper, thanks @lucaferranti for your work as editor and thank you all for your patience. 😄

We normally addressed all the unclarities/questions from your reviews. All the changes in the paper are written in red.

@jmskov

> - Currently, a couple of tests fail for me on MacOS on fresh Julia 1.6 and 1.10 in different ways. On 1.6, lines 213 and 214 of test_ellipsoids_abstraction.jl fail the approximation check, with Expression: ≈(cost_bound, 0.6250139513432214, rtol = 0.001). A MacOS+RNG issue? On 1.10, there is a conflict with radius with LazySets.jl. Perhaps a version issue.
> - Running benchmark.jl, I get an undefined SCOTSAbstraction, which I can't seem to find in the relevant module.

Thanks for reporting this! The two problems are normally solved in the latest version of Dionysos. Note that we bumped the Julia version of Dionysos to 1.10, as it will soon become the new LTS (as soon as 1.11 is out).

> - No explicit "Statement of need" section in the paper

We understood from previous reviews that "the consensus was that as long as the statement of need is clearly conveyed in the paper, it's not needed to have a section verbatim with that name" (citing @lucaferranti). Please let us know if you think that it would be better with an explicit "Statement of need" section.

> - For the problem definition, time::T, is the number of allowed steps. Can it handle a duration input that determines the maximum number of discrete steps?

If I understood well, yes! This parameter can typically be used like this.

> - Perhaps outside the scope of this package, but adding the capability to return a cost or another measure (probability eh?) with the controller would be useful for the future

Thanks for proposing new features! In the current state of Dionysos, in some situations, this cost can be computed a posteriori. We modified the caption of Figure 6 to include an example of computation of such a bound on the cost.

@schillic

> - There is a mismatch between Definition 3, defined over infinite words, and the behavior B(S) , which can be finite or infinite. Since specifications require only infinite behavior, but every infinite behavior induces finite prefixes, that makes any behavior violate any specification.

Thanks for reporting this problem. We modified Definition 3 so that both finite and infinite time specifications are handled.

> **Minor comments**

Thanks for reporting the unclarities and typos written in this section! We agree with all of the remarks, and we modified the paper accordingly to all of them. We answer to the following remark:

> - Fig. 2 and p. 3: I do not see why a strict single-valued map must induce a partition. Could R1 not map all states a to the same state b?

A partition is defined as a collection of sets that 1) covers the state space and 2) don't intersect with each other. A function (that is, a single-valued map) maps every point of the state space (making it a covering) to a unique symbol (making it non-intersecting with each other), and thereby automatically defines a partition. The case where R1 maps all states to the same state b corresponds to a trivial partition with only one bloc.

> #### Already reported, but for completeness

All problems reported in this section are normally solved in the latest version of Dionysos (see answer above).

> #### Further warnings
>
>  - `WARNING: using LazySets.vertices in module Utils conflicts with an existing identifier.`
>  - There are some duplicate method definitions (presumably from a merge conflict) in `src/optim/abstraction/`:

Thanks for spotting these two errors! Again, normally this has been solved in the latest version of Dionysos.
