### 2024-09-10 #160 @lucaferranti
> We understood from previous reviews that "the consensus was that as long as the statement of need is clearly conveyed in the paper, it's not needed to have a section verbatim with that name" (citing @lucaferranti). Please let us know if you think that it would be better with an explicit "Statement of need" section

correct! The work should of course be clearly motivated in the paper, but a section called statement of need is not needed. This was the most frequently asked question and we recently clarified that in the review checklist (unfortunately already generated checklists are not modififed). The new wording is:

> Does the paper clearly explain the problem it addresses, its importance to the broader community, the intended audience, and how it connects to existing work in the field?
