### 2024-10-07 7b8d67c
Refactor interface of Gol, Lazar, Belta
