### 2024-10-07 ab3ba1d
Add ideal notation
