### 2024-10-08 41d5351
dcdc example refactoring
