+++ docs/src/examples/solvers/DC-DC converter.jl
@@ -46,25 +46,23 @@ include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "dc_dc.jl"))
 # and we can instantiate the DC system with the provided system
 problem = DCDC.problem(; approx_mode = DCDC.GROWTH)
 
-controller = problem.solve(
+controller = problem.solve(;
     method = :uniform_abstraction,
-    initial_state = [0.0, 0.0], 
-    initial_input = [1.], 
+    initial_state = [0.0, 0.0],
+    initial_input = [1.0],
 )
 # Didn't work because the default step sizes are too coarse / too small
 problem.set_state_stepsize([2.0 / 4.0e3, 2.0 / 4.0e3])
-problem.set_input_stepsize([1.])
+problem.set_input_stepsize([1.0])
 controller = problem.solve()
 # Now it worked
-
 
 # ### Trajectory display
 # We choose the number of steps `nsteps` for the sampled system, i.e. the total elapsed time: `nstep`*`tstep`
 # as well as the true initial state `x0` which is contained in the initial state-space defined previously.
 nstep = 300
 x0 = [1.2, 5.6]
-control_trajectory =
-    ST.get_closed_loop_trajectory(problem.system.f, controller, x0, nstep)
+control_trajectory = ST.get_closed_loop_trajectory(problem.system.f, controller, x0, nstep)
 
 fig = plot(; aspect_ratio = :equal);
 plot!(problem.system.X);
@@ -86,12 +84,17 @@ P = SMatrix{2, 2}(1.0224, 0.0084, 0.0084, 1.0031)
 state_grid = DO.GridEllipsoidalRectangular(origin, SVector(η, η), P / ϵ, concrete_system.X)
 
 # Here we give a more complex grid to the solver
-controller = problem.solve(
-    method = :uniform_abstraction, 
-    δ_gas = true, 
-    state_grid = DO.GridEllipsoidalRectangular(origin, SVector(η, η), P / ϵ, concrete_system.X), 
-    initial_input = [1.], 
-    input_stepsize = [1.]
+controller = problem.solve(;
+    method = :uniform_abstraction,
+    δ_gas = true,
+    state_grid = DO.GridEllipsoidalRectangular(
+        origin,
+        SVector(η, η),
+        P / ϵ,
+        concrete_system.X,
+    ),
+    initial_input = [1.0],
+    input_stepsize = [1.0],
 )
 
 # ### Trajectory display
@@ -99,8 +102,7 @@ controller = problem.solve(
 # as well as the true initial state `x0` which is contained in the initial state-space defined previously.
 nstep = 300
 x0 = [1.2, 5.6]
-control_trajectory =
-    ST.get_closed_loop_trajectory(problem.system.f, controller, x0, nstep)
+control_trajectory = ST.get_closed_loop_trajectory(problem.system.f, controller, x0, nstep)
 
 fig = plot(; aspect_ratio = :equal);
 plot!(problem.system.X);
