+++ dsl/MACROS/problems/unicyclerobot.jl
@@ -1,0 +1,36 @@
+# Define Model with grid type
+model = Control("Unicycle Robot", UniformGrid())
+
+# Define state variables for the position and orientation
+x1 = @variable(model, "x1", Reals())
+x2 = @variable(model, "x2", Reals())
+x3 = @variable(model, "x3", Reals())
+
+# Define control inputs with bounds
+u1 = @variable(model, "u1", Reals(), (0.2, 2))
+u2 = @variable(model, "u2", Reals(), (-1, 1))
+
+# Define dynamics with modular arithmetic
+f_x1 = :(x1 + u1 * cos(x3))
+f_x2 = :(x2 + u1 * sin(x3))
+f_x3 = :(mod(x3 + u2, 2π))
+
+@constraint(model, :($f_x1 == x1))
+@constraint(model, :($f_x2 == x2))
+@constraint(model, :($f_x3 == x3))
+
+# Define set constraints
+@constraint(model, SetConstraint(:(x1^2 - x2^2 <= 4)))
+@constraint(model, SetConstraint(:(4 * x2^2 - x1^2 <= 16)))
+
+# Define stage and terminal costs
+stage_cost = StageCost(:(100 * ((x1 - x_r[1])^2 + (x2 - x_r[2])^2) + (u1^2 + u2^2)))
+terminal_cost = TerminalCost(:(100 * ((x1 - x_r[1])^2 + (x2 - x_r[2])^2)))
+
+# Set up MPC solution with horizon N = 20
+solution = model.solve_mpc([stage_cost.expression], terminal_cost.expression, horizon=20)
+
+# Visualize for reference positions
+for x_r in [(0.5, 0.5), (sqrt(32)/3, sqrt(20)/3)]
+    Visualize(solution, plot="state_trajectory", labels=["x1", "x2", "x3"], target=Target(x_r[1], x_r[2]))
+end
