### 2024-10-20 #160 @schillic
I confirm that all comments have been addressed. I have some new minor comments below, but they are optional.
As I already said before, a really nice piece of software and accompanying tool paper!

#### Minor comments regarding the paper

- Def. 6 and Def. 7: I suggest to put the formulas in centered math mode (`\[ ... \]`) for better readability.
- Example 1: "defined in 7" -> "defined in Definition 7"?

#### Minor comments regarding the package

- I sent a pull request to update some links in the README.
- FYI: In the [example from the documentation](https://dionysos-dev.github.io/Dionysos.jl/dev/generated/Path%20planning/#Definition-of-the-abstraction), an intermediate output of type `Dionysos.Utils.SortedTupleSet` cannot be printed (I tried it in Julia v1.11.1). You get a crash if you manually copy the code from the tutorial to the REPL. The code itself works fine if you add a semicolon. Here is the error message:
```
julia> abstract_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_controller"))

Error showing value of type Dionysos.Utils.SortedTupleSet{2, Tuple{Int64, Int64}}:
ERROR: MethodError: no method matching iterate(::Dionysos.Utils.SortedTupleSet{2, Tuple{Int64, Int64}})
The function `iterate` exists, but no method is defined for this combination of argument types.
```
- When running the tests, I see the following warnings:
   - `WARNING: method definition for system at Dionysos/problems/simple_problem.jl:28 declares type variable T but does not use it.`
   - `WARNING: method definition for system at Dionysos/problems/simple_problem.jl:28 declares type variable N but does not use it.`
