+++ Project.toml
@@ -41,7 +41,7 @@ Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 
 [compat]
 ArnoldiMethod = "0.2, 0.4"
-Colors = "0.12"
+Colors = "0.12, 0.13"
 DataStructures = "0.18"
 DiscreteMarkovChains = "0.2"
 Distributions = "0.25"
