### 2024-10-28 c570ea5
build abstraction from data done
