+++ src/optim/abstraction/sample_based_abstraction.jl
@@ -1,0 +1,305 @@
+export SampleBasedAbstraction
+
+module SampleBasedAbstraction
+
+import Dionysos
+const DI = Dionysos
+const UT = DI.Utils
+const DO = DI.Domain
+const ST = DI.System
+const SY = DI.Symbolic
+const PR = DI.Problem
+
+using JuMP
+
+"""
+    Optimizer{T} <: MOI.AbstractOptimizer
+
+TODO
+"""
+mutable struct Optimizer{T} <: MOI.AbstractOptimizer
+    concrete_problem::Union{Nothing, PR.ProblemType}
+    abstract_problem::Union{Nothing, PR.OptimalControlProblem, PR.SafetyProblem}
+    abstract_system::Union{Nothing, SY.SymbolicModelList}
+    abstract_controller::Union{Nothing, UT.SortedTupleSet{2, NTuple{2, Int}}}
+    concrete_controller::Any
+    state_grid::Union{Nothing, DO.Grid}
+    input_grid::Union{Nothing, DO.Grid}
+    solve_time_sec::T
+    function Optimizer{T}() where {T}
+        return new{T}(nothing, nothing, nothing, nothing, nothing, nothing, nothing, 0.0)
+    end
+end
+Optimizer() = Optimizer{Float64}()
+
+MOI.is_empty(optimizer::Optimizer) = optimizer.concrete_problem === nothing
+
+function MOI.set(model::Optimizer, param::MOI.RawOptimizerAttribute, value)
+    return setproperty!(model, Symbol(param.name), value)
+end
+function MOI.get(model::Optimizer, ::MOI.SolveTimeSec)
+    return model.solve_time_sec
+end
+function MOI.get(model::Optimizer, param::MOI.RawOptimizerAttribute)
+    return getproperty(model, Symbol(param.name))
+end
+
+function build_abstraction(concrete_system, state_grid::DO.Grid, input_grid::DO.Grid)
+    Xfull = DO.DomainList(state_grid)
+    DO.add_set!(Xfull, concrete_system.X, DO.INNER)
+    Ufull = DO.DomainList(input_grid)
+    DO.add_set!(Ufull, concrete_system.U, DO.CENTER)
+    abstract_system = SY.NewSymbolicModelListList(Xfull, Ufull)
+    @time SY.compute_symmodel_from_data!(abstract_system, concrete_system.f)
+    return abstract_system
+end
+
+
+
+
+
+function build_abstract_problem(
+    concrete_problem::PR.OptimalControlProblem,
+    abstract_system::SY.SymbolicModelList,
+)
+    state_grid = abstract_system.Xdom.grid
+    Xinit = DO.DomainList(state_grid)
+    DO.add_subset!(Xinit, abstract_system.Xdom, concrete_problem.initial_set, DO.OUTER)
+    Xtarget = DO.DomainList(state_grid)
+    DO.add_subset!(Xtarget, abstract_system.Xdom, concrete_problem.target_set, DO.INNER)
+    init_list = [SY.get_state_by_xpos(abstract_system, pos) for pos in DO.enum_pos(Xinit)]
+    target_list =
+        [SY.get_state_by_xpos(abstract_system, pos) for pos in DO.enum_pos(Xtarget)]
+    return PR.OptimalControlProblem(
+        abstract_system,
+        init_list,
+        target_list,
+        concrete_problem.state_cost, # TODO this is the continuous cost, not the abstraction
+        concrete_problem.transition_cost, # TODO this is the continuous cost, not the abstraction
+        concrete_problem.time, # TODO this is the continuous time, not the number of transition
+    )
+end
+
+function build_abstract_problem(
[…]
