### 2024-10-28 44d6877
revert DC example
