### 2024-10-28 db39e52
adding dc_dc_as_switch
