### 2024-10-28 d5f4242
adding more examples
