+++ dionysos-main/examples/planning_path.jl
@@ -1,42 +1,80 @@
-using InfiniteOpt, JuMP, DifferentialEquations, LinearAlgebra
+using CoreControls
+using Solver
 
-# Create an InfiniteOpt model
-model = InfiniteModel()
+# Create a Control model for the vehicle
+model = Control(
+    name="VehicleMazeNavigation",
+    problem_type=Reachability(),
+    system_type=Continuous()  # Continuous system type
+)
 
-# Define the time horizon (0 to T)
-@variable(model, t ∈ [0, T])
+# Parameters
+@parameter(model, "u_min", -1.0)
+@parameter(model, "u_max", 1.0)
+@parameter(model, "alpha", 0.0)  # Initialize alpha, computed later
 
-# Define the state variables: x1(t), x2(t), x3(t)
-@infinite_variable(model, x1(t))
-@infinite_variable(model, x2(t))
-@infinite_variable(model, x3(t))
+# State Variables (x1, x2 for position and x3 for orientation)
+@variable(model, "x1", StateVar(), Reals())
+@variable(model, "x2", StateVar(), Reals())
+@variable(model, "x3", StateVar(), Reals())
 
-# Define the control variables: u1(t), u2(t)
-@infinite_variable(model, -1 <= u1(t) <= 1)
-@infinite_variable(model, -1 <= u2(t) <= 1)
+# Control Inputs (u1 for velocity and u2 for steering angle)
+@variable(model, "u1", InputVar(), Reals(), bounds=(u_min, u_max))
+@variable(model, "u2", InputVar(), Reals(), bounds=(u_min, u_max))
 
-# Define the auxiliary variable for alpha
-@infinite_variable(model, α(t))
+# Dynamics for each state variable
+@constraint(model, :(dot(x1, t, Δt, model) == u1 * cos(alpha + x3) * cos(alpha)^-1))
+@constraint(model, :(dot(x2, t, Δt, model) == u1 * sin(alpha + x3) * cos(alpha)^-1))
+@constraint(model, :(dot(x3, t, Δt, model) == u1 * tan(u2)))
 
-# Set α(t) = arctan(tan(u2(t)) / 2)
-@infinite_constraint(model, α(t) == atan(tan(u2(t)) / 2))
+# Alpha is dependent on u2
+@constraint(model, :(alpha == atan(tan(u2) / 2)))
 
-# Define the dynamics of the system using the given equations
-@infinite_derivative(model, D(x1(t)) == u1(t) * cos(α(t) + x3(t)) / cos(α(t)))
-@infinite_derivative(model, D(x2(t)) == u1(t) * sin(α(t) + x3(t)) / cos(α(t)))
-@infinite_derivative(model, D(x3(t)) == u1(t) * tan(u2(t)))
+# Initial Conditions (Specify vehicle's starting position and orientation)
+x1_start = 0.0
+x2_start = 0.0
+x3_start = 0.0
+@constraint(model, :(x1_0 == $x1_start))
+@constraint(model, :(x2_0 == $x2_start))
+@constraint(model, :(x3_0 == $x3_start))
 
-# Initial conditions for the states
-@infinite_constraint(model, x1(0) == x1_initial)
-@infinite_constraint(model, x2(0) == x2_initial)
-@infinite_constraint(model, x3(0) == x3_initial)
+# Terminal Conditions (Target position)
+x1_goal = 10.0  # Example target position
+x2_goal = 10.0
+x3_goal = 0.0
+@constraint(model, :(x1_$N == $x1_goal))
[…]
