### 2024-10-28 3c21020
add simple pendulum
