### 2024-10-29 a1f2dad
adding examples and tweaking
