+++ Project.toml
@@ -16,6 +16,7 @@ HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
 HybridSystems = "2207ec0c-686c-5054-b4d2-543502888820"
 IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 IntervalLinearAlgebra = "92cbe1ac-9c24-436b-b0c9-5f7317aedcd5"
+Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LazySets = "b4f0291d-fe17-52bc-9479-3d1a343d9043"
@@ -52,6 +53,7 @@ HiGHS = "1"
 HybridSystems = "0.4"
 IntervalArithmetic = "0.20, 0.21, 0.22"
 IntervalLinearAlgebra = "0.1"
+Ipopt = "1.6.7"
 JuMP = "1"
 LaTeXStrings = "1"
 LazySets = "2"

[…]
