+++ Project.toml
@@ -24,6 +24,7 @@ LightXML = "9c8b4983-aa76-5018-a973-4c85ecc9e179"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
 MathematicalSystems = "d14a8603-c872-5ed3-9ece-53e0e82e39da"
+MeshCatMechanisms = "6ad125db-dd91-5488-b820-c1df6aab299d"
 MutableArithmetics = "d8a4904e-b15c-11e9-3269-09a3773c0cb0"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
