### 2024-10-29 3c074b0
step 1 for visualization
