### 2024-10-29 713513a
add analysis plots, add to the examples, and add sample based examples
