### 2024-10-29 fa64d61
adding thermostat examples
