### 2024-10-29 218c33d
changing imply operator
