+++ dionysos-main/examples/planning_path.jl
@@ -1,51 +1,47 @@
 # Import necessary modules
 include(joinpath(@__DIR__, "../src/core/core.jl"))
 
-using .CoreControls
+#using .CoreControls
 
 # Create a Control model for the vehicle
-model = @control(
+model = CoreControls.@control(
     name="VehicleMazeNavigation",
-    problem_type=Reachability(),
-    system_type=Continuous()  # Continuous system type
+    problem_type=CoreControls.Reachability(),
+    system_type=CoreControls.Continuous(),  # Continuous system type
 )
 
 # Parameters
-@parameter(model, "u_min", -1.0)
-@parameter(model, "u_max", 1.0)
-@parameter(model, "alpha", 0.0)  # Initialize alpha, computed later
-@parameter(model, "N", 10)  # Time step
+CoreControls.@parameter(model, "u_min", -1.0)
+CoreControls.@parameter(model, "u_max", 1.0)
+CoreControls.@parameter(model, "alpha", 0.0)  # Initialize alpha, computed later
+CoreControls.@parameter(model, "N", 10)  # Time step
 
 # State Variables (x1, x2 for position and x3 for orientation)
-@statevar(model, "x1", Reals())
-@statevar(model, "x2", Reals())
-@statevar(model, "x3", Reals())
+CoreControls.@statevar(model, "x1", CoreControls.Reals())
+CoreControls.@statevar(model, "x2", CoreControls.Reals())
+CoreControls.@statevar(model, "x3", CoreControls.Reals())
 
 # Control Inputs (u1 for velocity and u2 for steering angle)
-@inputvar(model, "u1", Reals(), bounds=(u_min, u_max))
-@inputvar(model, "u2", Reals(), bounds=(u_min, u_max))
+CoreControls.@inputvar(model, "u1", CoreControls.Reals(), bounds=(u_min, u_max))
+CoreControls.@inputvar(model, "u2", CoreControls.Reals(), bounds=(u_min, u_max))
 
 # Alpha is dependent on u2
-@constraint(model, alpha == atan(tan(u2) / 2))
-@constraint(model, dot(x1, t, Δt, model) == u1 * cos(alpha + x3) * cos(alpha)^-1)
-@constraint(model, dot(x2, t, Δt, model) == u1 * sin(alpha + x3) * cos(alpha)^-1)
-@constraint(model, dot(x3, t, Δt, model) == u1 * tan(u2))
+CoreControls.@constraint(model, dot(x1, t, Δt, model) == u1 * cos(atan(tan(u2) / 2) + x3) * cos(atan(tan(u2) / 2))^-1)
+CoreControls.@constraint(model, dot(x2, t, Δt, model) == u1 * sin(atan(tan(u2) / 2) + x3) * cos(atan(tan(u2) / 2))^-1)
+CoreControls.@constraint(model, dot(x3, t, Δt, model) == u1 * tan(u2))
 
 # Initial Conditions (Specify vehicle's starting position and orientation)
-x1_start = 0.0
-x2_start = 0.0
-x3_start = 0.0
-@constraint(model, x1_0 == x1_start)
-@constraint(model, x2_0 == x2_start)
-@constraint(model, x3_0 == x3_start)
+CoreControls.@parameter(model, "x1_start", 0.0)
+CoreControls.@parameter(model, "x2_start", 0.0)
+CoreControls.@parameter(model, "x3_start", 0.0)
+CoreControls.@constraint(model, x1_0 == x1_start)
+CoreControls.@constraint(model, x2_0 == x2_start)
[…]

[…]
