### 2024-10-30 7694538
Add JuMP interface
