+++ src/Dionysos.jl
@@ -1,5 +1,6 @@
 module Dionysos
 
+include("MOI_wrapper.jl")
 include("utils/utils.jl")
 include("domain/domain.jl")
 include("system/system.jl")
