+++ docs/src/examples/solvers/Path planning.jl
@@ -1,7 +1,46 @@
-using Test #src
-using StaticArrays, Dionysos, JuMP
+using Test     #src
+# # Example: Path planning problem solved by [Uniform grid abstraction](https://github.com/dionysos-dev/Dionysos.jl/blob/master/docs/src/manual/manual.md#solvers).
+#
+#md # [![Binder](https://mybinder.org/badge_logo.svg)](@__BINDER_ROOT_URL__/generated/Path planning.ipynb)
+#md # [![nbviewer](https://img.shields.io/badge/show-nbviewer-579ACA.svg)](@__NBVIEWER_ROOT_URL__/generated/Path planning.ipynb)
+#
+# This example was borrowed from [1, IX. Examples, A] whose dynamics comes from the model given in [2, Ch. 2.4].
+# This is a **reachability problem** for a **continuous system**.
+#
+# Let us consider the 3-dimensional state space control system of the form
+# ```math
+# \dot{x} = f(x, u)
+# ```
+# with $f: \mathbb{R}^3 × U ↦ \mathbb{R}^3$ given by
+# ```math
+# f(x,(u_1,u_2)) = \begin{bmatrix} u_1 \cos(α+x_3)\cos(α)^{-1} \\ u_1 \sin(α+x_3)\cos(α)^{-1} \\ u_1 \tan(u_2)  \end{bmatrix}
+# ```
+# and with $U = [−1, 1] \times [−1, 1]$ and $α = \arctan(\tan(u_2)/2)$. Here, $(x_1, x_2)$ is the position and $x_3$ is the
+# orientation of the vehicle in the 2-dimensional plane. The control inputs $u_1$ and $u_2$ are the rear
+# wheel velocity and the steering angle.
+# The control objective is to drive the vehicle which is situated in a maze made of obstacles from an initial position to a target position.
+#
+#
+# In order to study the concrete system and its symbolic abstraction in a unified framework, we will solve the problem
+# for the sampled system with a sampling time $\tau$.
+# For the construction of the relations in the abstraction, it is necessary to over-approximate attainable sets of
+# a particular cell. In this example, we consider the used of a growth bound function  [1, VIII.2, VIII.5] which is one of the possible methods to over-approximate
+# attainable sets of a particular cell based on the state reach by its center.
+#
+# For this reachability problem, the abstraction controller is built by solving a fixed-point equation which consists in computing the pre-image
+# of the target set.
 
-# Create an InfiniteOpt model
+# First, let us import [StaticArrays](https://github.com/JuliaArrays/StaticArrays.jl) and [Plots](https://github.com/JuliaPlots/Plots.jl).
+using StaticArrays, Plots
+
+# At this point, we import Dionysos and JuMP.
+using Dionysos, JuMP
+
+
+# ### Definition of the problem
+
+# We define the problem using JuMP as follows.
+# We first create a JuMP model:
 model = Model(Dionysos.Optimizer)
 
 # Define the state variables: x1(t), x2(t), x3(t)
@@ -25,7 +64,7 @@ x_target = [3.3, 0.5, 0]
 @constraint(model, final(x[2]) in MOI.Interval(0.3, 0.8))
 @constraint(model, final(x[3]) in MOI.Interval(-100.0, 100.0))
 
-# Obstacle boun daries (provided)
+# Obstacle boundaries (provided)
 x1_lb = [1.0, 2.2, 2.2]
 x1_ub = [1.2, 2.4, 2.4]
 x2_lb = [0.0, 0.0, 6.0]
@@ -37,7 +76,9 @@ for i in eachindex(x1_ub)
[…]
