+++ src/MOI_wrapper.jl
@@ -6,9 +6,9 @@ import JuMP
 @enum(VariableType, INPUT, STATE, MODE)
 
 mutable struct Optimizer <: MOI.AbstractOptimizer
-    inner
-    nlp_model
-    evaluator::Union{Nothing,MOI.Nonlinear.Evaluator}
+    inner::Any
+    nlp_model::Any
+    evaluator::Union{Nothing, MOI.Nonlinear.Evaluator}
     variable_values::Vector{Float64}
     derivative_values::Vector{Float64}
     variable_types::Vector{VariableType}
@@ -17,8 +17,8 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
     upper::Vector{Float64}
     start::Vector{Float64}
     target::Vector{MOI.Interval{Float64}}
-    dynamic::Vector{Union{Nothing,MOI.ScalarNonlinearFunction}}
-    obstacles::Vector{Tuple{Vector{MOI.VariableIndex},MOI.HyperRectangle}}
+    dynamic::Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}
+    obstacles::Vector{Tuple{Vector{MOI.VariableIndex}, MOI.HyperRectangle}}
     objective_sense::MOI.OptimizationSense
     objective_function::MOI.AbstractScalarFunction
     nonlinear_index::Int
@@ -35,8 +35,8 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
             Float64[],
             Float64[],
             MOI.Interval{Float64}[],
-            Union{Nothing,MOI.ScalarNonlinearFunction}[],
-            Tuple{Vector{MOI.VariableIndex},MOI.HyperRectangle}[],
+            Union{Nothing, MOI.ScalarNonlinearFunction}[],
+            Tuple{Vector{MOI.VariableIndex}, MOI.HyperRectangle}[],
             MOI.FEASIBILITY_SENSE,
             zero(MOI.ScalarAffineFunction{Float64}),
             0,
@@ -75,7 +75,7 @@ function MOI.add_variable(model::Optimizer)
     return MOI.VariableIndex(length(model.upper))
 end
 
-struct OuterSet{S<:MOI.AbstractVectorSet} <: MOI.AbstractVectorSet
+struct OuterSet{S <: MOI.AbstractVectorSet} <: MOI.AbstractVectorSet
     inner::S
 end
 
@@ -85,7 +85,7 @@ Base.copy(set::OuterSet) = OuterSet(copy(set.inner))
 function MOI.supports_constraint(
     ::Optimizer,
     ::Type{MOI.VectorOfVariables},
-    ::Type{OuterSet{MOI.HyperRectangle{Float64}}}
+    ::Type{OuterSet{MOI.HyperRectangle{Float64}}},
 )
     return true
 end
@@ -96,33 +96,25 @@ function MOI.add_constraint(
     set::OuterSet{MOI.HyperRectangle{Float64}},
 )
     push!(model.obstacles, (func.variables, set.inner))
-    return MOI.ConstraintIndex{typeof(func),typeof(set)}(length(model.obstacles))
+    return MOI.ConstraintIndex{typeof(func), typeof(set)}(length(model.obstacles))
 end
 
 function MOI.supports_constraint(
     ::Optimizer,
     ::Type{MOI.VariableIndex},
-    ::Type{<:Union{MOI.LessThan,MOI.GreaterThan}}
+    ::Type{<:Union{MOI.LessThan, MOI.GreaterThan}},
 )
     return true
 end
 
-function MOI.add_constraint(
-    model::Optimizer,
-    func::MOI.VariableIndex,
-    set::MOI.GreaterThan
-)
[…]
