+++ .github/workflows/format_check.yml
@@ -13,8 +13,8 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        julia-version: [1.3.0]
-        julia-arch: [x86]
+        julia-version: [1.10.0]
+        julia-arch: [x64]
         os: [ubuntu-latest]
     steps:
       - uses: julia-actions/setup-julia@latest
@@ -41,4 +41,3 @@ jobs:
               exit(1)
           end'
       
-
