+++ Project.toml
@@ -23,6 +23,7 @@ Libtiff_jll = "89763e89-9b03-5906-acba-b20f662cd828"
 LightXML = "9c8b4983-aa76-5018-a973-4c85ecc9e179"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
+MathOptSymbolicAD = "309f4015-3481-4d63-a8f9-aeb13adfe8eb"
 MathematicalSystems = "d14a8603-c872-5ed3-9ece-53e0e82e39da"
 MutableArithmetics = "d8a4904e-b15c-11e9-3269-09a3773c0cb0"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
@@ -57,6 +58,7 @@ LazySets = "2"
 Libtiff_jll = "4.4.0"
 LightXML = "0.9"
 MathOptInterface = "1"
+MathOptSymbolicAD = "0.2.1"
 MathematicalSystems = "0.11, 0.12, 0.13"
 MutableArithmetics = "1"
 Plots = "1"
