+++ docs/src/examples/solvers/Path planning.jl
@@ -36,7 +36,6 @@ using StaticArrays, Plots
 # At this point, we import Dionysos and JuMP.
 using Dionysos, JuMP
 
-
 # ### Definition of the problem
 
 # We define the problem using JuMP as follows.
@@ -46,7 +45,7 @@ model = Model(Dionysos.Optimizer)
 # Define the state variables: x1(t), x2(t), x3(t)
 x_low, x_upp = [0.0, 0.0, -pi - 0.4], [4.0, 10.0, pi + 0.4]
 x_start = [0.4, 0.4, 0.0]
-@variable(model, x_low[i] <= x[i=1:3] <= x_upp[i], start = x_start[i])
+@variable(model, x_low[i] <= x[i = 1:3] <= x_upp[i], start = x_start[i])
 
 # Define the control variables: u1(t), u2(t)
 @variable(model, -1 <= u[1:2] <= 1)
@@ -73,7 +72,10 @@ x2_ub = [9.0, 5.0, 10.0]
 # Function to add rectangular obstacle avoidance constraints
 
 for i in eachindex(x1_ub)
-    @constraint(model, x[1:2] ∉ MOI.HyperRectangle([x1_lb[i], x2_lb[i]], [x1_ub[i], x2_ub[i]]))
+    @constraint(
+        model,
+        x[1:2] ∉ MOI.HyperRectangle([x1_lb[i], x2_lb[i]], [x1_ub[i], x2_ub[i]])
+    )
 end
 
 # ### Definition of the abstraction
@@ -138,7 +140,10 @@ plot!(concrete_problem.target_set; dims = [1, 2], color = :red, opacity = 0.2);
 
 # We display the abstract specifications
 plot!(
-    Dionysos.Symbolic.get_domain_from_symbols(abstract_system, abstract_problem.initial_set);
+    Dionysos.Symbolic.get_domain_from_symbols(
+        abstract_system,
+        abstract_problem.initial_set,
+    );
     color = :green,
 );
 plot!(
