### 2024-10-30 3cf2de1
Remove RigidBodyDynamics dependency (#377)
