### 2024-10-30 c4b1f35
clean simple pendulum
