+++ docs/src/examples/solvers/Pendulum/Safety.jl
@@ -1,7 +1,6 @@
-using Test     #src
+using Test  
 using StaticArrays, Plots
 
-# At this point, we import the useful Dionysos sub-modules.
 using Dionysos
 const DI = Dionysos
 const UT = DI.Utils
@@ -11,14 +10,13 @@ const SY = DI.Symbolic
 const OP = DI.Optim
 const AB = OP.Abstraction
 
-include(joinpath(dirname(dirname(pathof(Dionysos))), "problems/simple_pendulum", "safety_unstable_equilibrium.jl"))  # safety_stable_equilibrium.jl safety_unstable_equilibrium.jl
+include(joinpath(dirname(dirname(pathof(Dionysos))), "problems/pendulum", "pendulum.jl"))
 
-# and we can instantiate the DC system with the provided system
-concrete_problem = Pendulum.problem(; approx_mode = "growth")
+concrete_problem = Pendulum.problem(; objective = "safety-up")
 concrete_system = concrete_problem.system
 
 x0 = SVector(0.0, 0.0)
-hx = SVector(0.1, 0.1)
+hx = SVector(0.05, 0.05)
 state_grid = DO.GridFree(x0, hx)
 
 u0 = SVector(0.0);
@@ -36,13 +34,27 @@ abstract_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_con
 concrete_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("concrete_controller"))
 
 # ### Trajectory display
-# We choose the number of steps `nsteps` for the sampled system, i.e. the total elapsed time: `nstep`*`tstep`
-# as well as the true initial state `x0` which is contained in the initial state-space defined previously.
 nstep = 100
-x0 = SVector(pi+0.15,0.5) # SVector(0.15,0.0) #
+x0 = SVector(UT.sample(concrete_problem.initial_set)...)
+
 control_trajectory =
     ST.get_closed_loop_trajectory(concrete_system.f, concrete_controller, x0, nstep)
 
 fig = plot(; aspect_ratio = :equal);
 plot!(concrete_system.X);
+plot!(concrete_problem.safe_set; markersize=1,arrows=false)
 plot!(control_trajectory; markersize=1,arrows=false)
+display(fig)
+
+# ### For Visualization
+urdf = joinpath(dirname(dirname(pathof(Dionysos))), "problems/pendulum/", "Pendulum.urdf")
+tstep = 0.1
+state_values = [ST.get_state(control_trajectory, i)[1] for i in 1:ST.length(control_trajectory)]
+ts = collect(0.0:tstep:((length(state_values)-1) * tstep))
+qs = Vector{Vector{Float64}}(undef, length(state_values))
+for i in 1:length(state_values)
+    set_configuration!(state, shoulder, state_values[i])
+    qs[i] = configuration(state)
+end
+mvis = MechanismVisualizer(mechanism, URDFVisuals(urdf))
+MeshCatMechanisms.setanimation!(mvis, ts, qs)
