### 2024-10-30 dd36535
clean simple pendulum
