+++ Project.toml
@@ -32,6 +32,7 @@ Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+RigidBodyDynamics = "366cf18f-59d5-5db9-a4de-86a9f6786172"
 SemialgebraicSets = "8e049039-38e8-557d-ae3a-bc521ccf6204"
 SimpleWeightedGraphs = "47aef6b3-ad0c-573a-a1e2-d07658019622"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
