+++ docs/src/examples/solvers/Pendulum/Safety.jl
@@ -47,13 +47,18 @@ plot!(control_trajectory; markersize=1,arrows=false)
 display(fig)
 
 # ### For Visualization
+using RigidBodyDynamics, MeshCatMechanisms
 urdf = joinpath(dirname(dirname(pathof(Dionysos))), "problems/pendulum/", "Pendulum.urdf")
+mechanism = parse_urdf(urdf)
+state = MechanismState(mechanism)
+joint = first(joints(mechanism))
+
 tstep = 0.1
 state_values = [ST.get_state(control_trajectory, i)[1] for i in 1:ST.length(control_trajectory)]
 ts = collect(0.0:tstep:((length(state_values)-1) * tstep))
 qs = Vector{Vector{Float64}}(undef, length(state_values))
 for i in 1:length(state_values)
-    set_configuration!(state, shoulder, state_values[i])
+    set_configuration!(state, joint, state_values[i])
     qs[i] = configuration(state)
 end
 mvis = MechanismVisualizer(mechanism, URDFVisuals(urdf))
