+++ src/MOI_wrapper.jl
@@ -261,7 +261,13 @@ function dynamicofsystem(model, x_idx, u_idx)
     # `cse = true` is used to detect common sub-expressions (like `α` in the path planning example),
     # cfr. https://discourse.julialang.org/t/detecting-function-composition-in-symbolics-jl/115885/6
     # for a discussion on why it's not currently enabled by default
-    F_sys, _ = Symbolics.build_function(expr, collect(x), collect(u), expression = Val{false}, cse = true)
+    F_sys, _ = Symbolics.build_function(
+        expr,
+        collect(x),
+        collect(u);
+        expression = Val{false},
+        cse = true,
+    )
 
     # To see the generated expression, use:
     #println(Symbolics.build_function(expr, collect(x), collect(u), cse = true))
