### 2024-10-30 ade4958
Add JuMP interface (#376)
* Add JuMP interface

* up

* Fix format

* Update format check

* Fixes

* up versions in ci

* Fix format

* Fix perf issue

* Fix format
