+++ docs/src/examples/solvers/Path planning.jl
@@ -33,52 +33,73 @@ using Test     #src
 # First, let us import [StaticArrays](https://github.com/JuliaArrays/StaticArrays.jl) and [Plots](https://github.com/JuliaPlots/Plots.jl).
 using StaticArrays, Plots
 
-# At this point, we import Dionysos.
-using Dionysos
-const DI = Dionysos
-const UT = DI.Utils
-const DO = DI.Domain
-const ST = DI.System
-const SY = DI.Symbolic
-const PR = DI.Problem
-const OP = DI.Optim
-const AB = OP.Abstraction
-
-# And the file defining the hybrid system for this problem
-include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "path_planning.jl"))
+# At this point, we import Dionysos and JuMP.
+using Dionysos, JuMP
 
 # ### Definition of the problem
 
-# Now we instantiate the problem using the function provided by [PathPlanning.jl](@__REPO_ROOT_URL__/problems/PathPlanning.jl) 
-concrete_problem = PathPlanning.problem(; simple = true, approx_mode = PathPlanning.GROWTH);
-concrete_system = concrete_problem.system;
+# We define the problem using JuMP as follows.
+# We first create a JuMP model:
+model = Model(Dionysos.Optimizer)
+
+# Define the state variables: x1(t), x2(t), x3(t)
+x_low, x_upp = [0.0, 0.0, -pi - 0.4], [4.0, 10.0, pi + 0.4]
+x_start = [0.4, 0.4, 0.0]
+@variable(model, x_low[i] <= x[i = 1:3] <= x_upp[i], start = x_start[i])
+
+# Define the control variables: u1(t), u2(t)
+@variable(model, -1 <= u[1:2] <= 1)
+
+# Set α(t) = arctan(tan(u2(t)) / 2)
+@expression(model, α, atan(tan(u[2]) / 2))
+
+@constraint(model, ∂(x[1]) == u[1] * cos(α + x[3]) * sec(α))
+@constraint(model, ∂(x[2]) == u[1] * sin(α + x[3]) * sec(α))
+@constraint(model, ∂(x[3]) == u[1] * tan(u[2]))
+
+x_target = [3.3, 0.5, 0]
+
+@constraint(model, final(x[1]) in MOI.Interval(3.0, 3.6))
+@constraint(model, final(x[2]) in MOI.Interval(0.3, 0.8))
+@constraint(model, final(x[3]) in MOI.Interval(-100.0, 100.0))
+
+# Obstacle boundaries (provided)
+x1_lb = [1.0, 2.2, 2.2]
+x1_ub = [1.2, 2.4, 2.4]
+x2_lb = [0.0, 0.0, 6.0]
+x2_ub = [9.0, 5.0, 10.0]
+
+# Function to add rectangular obstacle avoidance constraints
+
+for i in eachindex(x1_ub)
+    @constraint(
+        model,
+        x[1:2] ∉ MOI.HyperRectangle([x1_lb[i], x2_lb[i]], [x1_ub[i], x2_ub[i]])
+    )
+end
 
 # ### Definition of the abstraction
 
 # Definition of the grid of the state-space on which the abstraction is based (origin `x0` and state-space discretization `h`):
+
 x0 = SVector(0.0, 0.0, 0.0);
 h = SVector(0.2, 0.2, 0.2);
-state_grid = DO.GridFree(x0, h);
+set_attribute(model, "state_grid", Dionysos.Domain.GridFree(x0, h))
 
 # Definition of the grid of the input-space on which the abstraction is based (origin `u0` and input-space discretization `h`):
 u0 = SVector(0.0, 0.0);
[…]
