+++ src/MOI_wrapper.jl
@@ -1,0 +1,432 @@
+import MathOptInterface as MOI
+import StaticArrays as SA
+import MathematicalSystems
+import JuMP
+import MathOptSymbolicAD
+import Symbolics
+
+@enum(VariableType, INPUT, STATE, MODE)
+
+mutable struct Optimizer <: MOI.AbstractOptimizer
+    inner::Any
+    nlp_model::Any
+    evaluator::Union{Nothing, MOI.Nonlinear.Evaluator}
+    variable_values::Vector{Float64}
+    derivative_values::Vector{Float64}
+    variable_types::Vector{VariableType}
+    variable_index::Vector{Int} # MOI.VariableIndex -> state/input/mode index
+    lower::Vector{Float64}
+    upper::Vector{Float64}
+    start::Vector{Float64}
+    target::Vector{MOI.Interval{Float64}}
+    dynamic::Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}
+    obstacles::Vector{Tuple{Vector{MOI.VariableIndex}, MOI.HyperRectangle}}
+    objective_sense::MOI.OptimizationSense
+    objective_function::MOI.AbstractScalarFunction
+    nonlinear_index::Int
+    function Optimizer()
+        return new(
+            MOI.instantiate(Dionysos.Optim.Abstraction.UniformGridAbstraction.Optimizer),
+            nothing,
+            nothing,
+            Float64[],
+            Float64[],
+            VariableType[],
+            Int[],
+            Float64[],
+            Float64[],
+            Float64[],
+            MOI.Interval{Float64}[],
+            Union{Nothing, MOI.ScalarNonlinearFunction}[],
+            Tuple{Vector{MOI.VariableIndex}, MOI.HyperRectangle}[],
+            MOI.FEASIBILITY_SENSE,
+            zero(MOI.ScalarAffineFunction{Float64}),
+            0,
+        )
+    end
+end
+
+MOI.is_empty(model::Optimizer) = isempty(model.variable_types)
+
+function MOI.empty!(model::Optimizer)
+    empty!(model.variable_values)
+    empty!(model.derivative_values)
+    empty!(model.variable_types)
+    empty!(model.variable_index)
+    empty!(model.lower)
+    empty!(model.upper)
+    empty!(model.start)
+    empty!(model.target)
+    empty!(model.dynamic)
+    empty!(model.obstacles)
+    model.objective_sense = MOI.FEASIBILITY_SENSE
+    model.objective_function = zero(MOI.ScalarAffineFunction{Float64})
+    model.nonlinear_index = 0
+    return
+end
+
+function MOI.add_variable(model::Optimizer)
+    push!(model.variable_values, NaN)
+    push!(model.variable_types, INPUT)
+    push!(model.variable_index, 0)
+    push!(model.lower, -Inf)
+    push!(model.upper, Inf)
+    push!(model.start, NaN)
+    push!(model.target, MOI.Interval(-Inf, Inf))
+    push!(model.dynamic, nothing)
+    return MOI.VariableIndex(length(model.upper))
+end
+
+struct OuterSet{S <: MOI.AbstractVectorSet} <: MOI.AbstractVectorSet
+    inner::S
+end
+
+MOI.dimension(set::OuterSet) = MOI.dimension(set.inner)
+Base.copy(set::OuterSet) = OuterSet(copy(set.inner))
+
+function MOI.supports_constraint(
+    ::Optimizer,
+    ::Type{MOI.VectorOfVariables},
+    ::Type{OuterSet{MOI.HyperRectangle{Float64}}},
+)
+    return true
+end
+
+function MOI.add_constraint(
+    model::Optimizer,
+    func::MOI.VectorOfVariables,
+    set::OuterSet{MOI.HyperRectangle{Float64}},
+)
[…]
