### 2024-10-30 5aee472
Move solver parameters to attributes
