### 2024-10-30 af70d36
Add test/Project.toml
