+++ test/Project.toml
@@ -1,0 +1,15 @@
+[deps]
+CDDLib = "3391f64e-dcde-5f30-b752-e11513730f60"
+Clarabel = "61c947e1-3e6d-4ee4-985a-eec8c727bd6e"
+Dionysos = "d92c97cf-b87d-42c1-a9c0-25df00b4d958"
+HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
+IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
+Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
+JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+Mosek = "6405355b-0ac2-5fba-af84-adbd65488c0e"
+MosekTools = "1ec41992-ff65-5c91-ac43-2df89e9693a4"
+OSQP = "ab2f91bb-94b4-55e3-9ba0-7f65df51de79"
+Pavito = "cd433a01-47d1-575d-afb7-6db927ee8d8f"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
