+++ Project.toml
@@ -71,17 +71,3 @@ StaticArrays = "1"
 Suppressor = "0.2"
 Symbolics = "5, 6"
 julia = "1.10"
-
-[extras]
-CDDLib = "3391f64e-dcde-5f30-b752-e11513730f60"
-Clarabel = "61c947e1-3e6d-4ee4-985a-eec8c727bd6e"
-HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
-Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
-Mosek = "6405355b-0ac2-5fba-af84-adbd65488c0e"
-MosekTools = "1ec41992-ff65-5c91-ac43-2df89e9693a4"
-OSQP = "ab2f91bb-94b4-55e3-9ba0-7f65df51de79"
-Pavito = "cd433a01-47d1-575d-afb7-6db927ee8d8f"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["CDDLib", "Ipopt", "OSQP", "Pavito", "Clarabel", "Test", "HiGHS", "Mosek", "MosekTools"]
