+++ dev/generated/Path planning.jl
@@ -1,39 +1,55 @@
 using StaticArrays, Plots
 
-using Dionysos
-const DI = Dionysos
-const UT = DI.Utils
-const DO = DI.Domain
-const ST = DI.System
-const SY = DI.Symbolic
-const PR = DI.Problem
-const OP = DI.Optim
-const AB = OP.Abstraction
+using Dionysos, JuMP
 
-include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "path_planning.jl"))
+model = Model(Dionysos.Optimizer)
 
-concrete_problem = PathPlanning.problem(; simple = true, approx_mode = PathPlanning.GROWTH);
-concrete_system = concrete_problem.system;
+x_low, x_upp = [0.0, 0.0, -pi - 0.4], [4.0, 10.0, pi + 0.4]
+x_start = [0.4, 0.4, 0.0]
+@variable(model, x_low[i] <= x[i = 1:3] <= x_upp[i], start = x_start[i])
+
+@variable(model, -1 <= u[1:2] <= 1)
+
+@expression(model, α, atan(tan(u[2]) / 2))
+
+@constraint(model, ∂(x[1]) == u[1] * cos(α + x[3]) * sec(α))
+@constraint(model, ∂(x[2]) == u[1] * sin(α + x[3]) * sec(α))
+@constraint(model, ∂(x[3]) == u[1] * tan(u[2]))
+
+x_target = [3.3, 0.5, 0]
+
+@constraint(model, final(x[1]) in MOI.Interval(3.0, 3.6))
+@constraint(model, final(x[2]) in MOI.Interval(0.3, 0.8))
+@constraint(model, final(x[3]) in MOI.Interval(-100.0, 100.0))
+
+x1_lb = [1.0, 2.2, 2.2]
+x1_ub = [1.2, 2.4, 2.4]
+x2_lb = [0.0, 0.0, 6.0]
+x2_ub = [9.0, 5.0, 10.0]
+
+for i in eachindex(x1_ub)
+    @constraint(
+        model,
+        x[1:2] ∉ MOI.HyperRectangle([x1_lb[i], x2_lb[i]], [x1_ub[i], x2_ub[i]])
+    )
+end
 
 x0 = SVector(0.0, 0.0, 0.0);
 h = SVector(0.2, 0.2, 0.2);
-state_grid = DO.GridFree(x0, h);
+set_attribute(model, "state_grid", Dionysos.Domain.GridFree(x0, h))
 
 u0 = SVector(0.0, 0.0);
 h = SVector(0.3, 0.3);
-input_grid = DO.GridFree(u0, h);
+set_attribute(model, "input_grid", Dionysos.Domain.GridFree(u0, h))
 
-using JuMP
-optimizer = MOI.instantiate(AB.UniformGridAbstraction.Optimizer)
-MOI.set(optimizer, MOI.RawOptimizerAttribute("concrete_problem"), concrete_problem)
-MOI.set(optimizer, MOI.RawOptimizerAttribute("state_grid"), state_grid)
-MOI.set(optimizer, MOI.RawOptimizerAttribute("input_grid"), input_grid)
-MOI.optimize!(optimizer)
+optimize!(model)
 
-abstract_system = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_system"))
-abstract_problem = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_problem"))
-abstract_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_controller"))
-concrete_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("concrete_controller"))
+abstract_system = get_attribute(model, "abstract_system");
[…]
