+++ previews/PR380/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"EditURL = \"../examples/Getting Started.jl\"","category":"page"},{"location":"generated/Getting Started/#Getting-Started","page":"Getting Started","title":"Getting Started","text":"","category":"section"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"(Image: Binder) (Image: nbviewer)","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"In this file we will visit the basic functionalities provided by Dionysos for the optimal control of complex systems. In summary, the topics covered are","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"Grids and discretizations\nDynamical system declaration\nContinuous and discrete state image mapping\nPlotting","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"First, let us import a few packages that are necessary to run this example.","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"using Dionysos\nusing StaticArrays\nusing LinearAlgebra\nusing Plots","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"The main package Dionysos provides most important data structures that we will need. Additionally  StaticArrays provides faster implementation of Arrays (which have static memory allocation), LinearAlgebra allows us to perform some additional operations.","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"const DI = Dionysos\nconst UT = DI.Utils\nconst DO = DI.Domain\nconst ST = DI.System\nconst SY = DI.Symbolic","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"Additionally, we will short the submodules accondingly","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"We use HyperRectangle to represent the boundary of the state space rectX and the input space rectU.","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"rectX = UT.HyperRectangle(SVector(-2, -2), SVector(2, 2));\nrectU = UT.HyperRectangle(SVector(-5), SVector(5));\nnothing #hide","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"A discretization of the state space is declared using the GridFree structure, which requires the definition of a center x0 and a vector h of discretization steps in each direction.","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"x0 = SVector(0.0, 0.0);\nh = SVector(1.0 / 5, 1.0 / 5);\nXgrid = DO.GridFree(x0, h);\nnothing #hide","category":"page"},{"location":"generated/Getting Started/","page":"Getting Started","title":"Getting Started","text":"Xgrid represents the state space grid and holds information of x0 and h, but is not a collection of cells. Indeed, a cell can be efficiently represented by a tuple of Int, for instance 'pos', with which the corresponding cartesian position can be computed by x0 + h .* pos or using functions to be shown. In Dionysos, a set of cells is called a Domain and the DomainList 
[…]
