### 2024-10-30 09006f8
Add Dionysos to docs project (#379)
