+++ docs/Project.toml
@@ -2,6 +2,7 @@
 CDDLib = "3391f64e-dcde-5f30-b752-e11513730f60"
 Clarabel = "61c947e1-3e6d-4ee4-985a-eec8c727bd6e"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
+Dionysos = "d92c97cf-b87d-42c1-a9c0-25df00b4d958"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
