+++ test/Project.toml
@@ -3,6 +3,7 @@ CDDLib = "3391f64e-dcde-5f30-b752-e11513730f60"
 Clarabel = "61c947e1-3e6d-4ee4-985a-eec8c727bd6e"
 Dionysos = "d92c97cf-b87d-42c1-a9c0-25df00b4d958"
 HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
+HybridSystems = "2207ec0c-686c-5054-b4d2-543502888820"
 IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
