+++ test/optim/test_UniformGridAbstraction_safety.jl
@@ -18,7 +18,7 @@ const AB = OP.Abstraction
 include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "dc_dc.jl"))
 
 # and we can instantiate the DC system with the provided system
-concrete_problem = DCDC.problem(; approx_mode = DCDC.GROWTH)
+concrete_problem = DCDC.problem()
 concrete_system = concrete_problem.system
 
 x0 = SVector(0.0, 0.0)
@@ -33,6 +33,9 @@ optimizer = MOI.instantiate(AB.UniformGridAbstraction.Optimizer)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("concrete_problem"), concrete_problem)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("state_grid"), state_grid)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("input_grid"), input_grid)
+MOI.set(optimizer, MOI.RawOptimizerAttribute("approx_mode"), Dionysos.Optim.Abstraction.UniformGridAbstraction.GROWTH)
+MOI.set(optimizer, MOI.RawOptimizerAttribute("jacobian_bound"), DCDC.jacobian_bound())
+MOI.set(optimizer, MOI.RawOptimizerAttribute("time_step"), 0.5)
 MOI.optimize!(optimizer)
 
 abstract_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_controller"))
