+++ test/Project.toml
@@ -15,4 +15,5 @@ OSQP = "ab2f91bb-94b4-55e3-9ba0-7f65df51de79"
 Pavito = "cd433a01-47d1-575d-afb7-6db927ee8d8f"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
