+++ test/Project.toml
@@ -1,6 +1,7 @@
 [deps]
 CDDLib = "3391f64e-dcde-5f30-b752-e11513730f60"
 Clarabel = "61c947e1-3e6d-4ee4-985a-eec8c727bd6e"
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 Dionysos = "d92c97cf-b87d-42c1-a9c0-25df00b4d958"
 HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
 HybridSystems = "2207ec0c-686c-5054-b4d2-543502888820"
@@ -16,6 +17,7 @@ Pavito = "cd433a01-47d1-575d-afb7-6db927ee8d8f"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+SemialgebraicSets = "8e049039-38e8-557d-ae3a-bc521ccf6204"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
