### 2024-10-30 b9aaf69
Add test/Project.toml (#380)
* Add test/Project.toml

* Add LinearAlgebra

* Add HS

* Add MathematicalSystems

* Add Symbolics

* Add Polyhedra

* Add Random

* up

* up
