+++ test/optim/test_UniformGridAbstraction_reachability.jl
@@ -20,7 +20,7 @@ include("../../problems/path_planning.jl")
 # ### Definition of the problem
 
 # Now we instantiate the problem using the function provided by [PathPlanning.jl](@__REPO_ROOT_URL__/problems/PathPlanning.jl) 
-concrete_problem = PathPlanning.problem(; simple = true, approx_mode = PathPlanning.GROWTH)
+concrete_problem = PathPlanning.problem(; simple = true)
 concrete_system = concrete_problem.system
 
 # ### Definition of the abstraction
@@ -42,6 +42,7 @@ optimizer = MOI.instantiate(AB.UniformGridAbstraction.Optimizer)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("concrete_problem"), concrete_problem)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("state_grid"), state_grid)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("input_grid"), input_grid)
+MOI.set(optimizer, MOI.RawOptimizerAttribute("jacobian_bound"), PathPlanning.jacobian_bound())
 MOI.optimize!(optimizer)
 
 # Get the results
