+++ previews/PR378/examples/solvers/DC-DC converter.jl
@@ -108,7 +108,11 @@ optimizer = MOI.instantiate(AB.UniformGridAbstraction.Optimizer)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("concrete_problem"), concrete_problem)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("state_grid"), state_grid)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("input_grid"), input_grid)
-MOI.set(optimizer, MOI.RawOptimizerAttribute("approx_mode"), Dionysos.Optim.Abstraction.UniformGridAbstraction.DELTA_GAS)
+MOI.set(
+    optimizer,
+    MOI.RawOptimizerAttribute("approx_mode"),
+    Dionysos.Optim.Abstraction.UniformGridAbstraction.DELTA_GAS,
+)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("δGAS"), true)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("time_step"), 0.5)
 MOI.optimize!(optimizer)
