+++ docs/src/examples/solvers/Pendulum/Safety.jl
@@ -1,4 +1,4 @@
-using Test  
+using Test
 using StaticArrays, Plots
 
 using Dionysos
@@ -42,8 +42,8 @@ control_trajectory =
 
 fig = plot(; aspect_ratio = :equal);
 plot!(concrete_system.X);
-plot!(concrete_problem.safe_set; markersize=1,arrows=false)
-plot!(control_trajectory; markersize=1,arrows=false)
+plot!(concrete_problem.safe_set; markersize = 1, arrows = false)
+plot!(control_trajectory; markersize = 1, arrows = false)
 display(fig)
 
 # ### For Visualization
@@ -54,8 +54,9 @@ state = MechanismState(mechanism)
 joint = first(joints(mechanism))
 
 tstep = 0.1
-state_values = [ST.get_state(control_trajectory, i)[1] for i in 1:ST.length(control_trajectory)]
-ts = collect(0.0:tstep:((length(state_values)-1) * tstep))
+state_values =
+    [ST.get_state(control_trajectory, i)[1] for i in 1:ST.length(control_trajectory)]
+ts = collect(0.0:tstep:((length(state_values) - 1) * tstep))
 qs = Vector{Vector{Float64}}(undef, length(state_values))
 for i in 1:length(state_values)
     set_configuration!(state, joint, state_values[i])
