+++ docs/Project.toml
@@ -17,6 +17,7 @@ Libtiff_jll = "89763e89-9b03-5906-acba-b20f662cd828"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 MathematicalSystems = "d14a8603-c872-5ed3-9ece-53e0e82e39da"
+MeshCatMechanisms = "6ad125db-dd91-5488-b820-c1df6aab299d"
 Mosek = "6405355b-0ac2-5fba-af84-adbd65488c0e"
 MosekTools = "1ec41992-ff65-5c91-ac43-2df89e9693a4"
 OSQP = "ab2f91bb-94b4-55e3-9ba0-7f65df51de79"
