### 2024-10-31 6f36279
Delete double pendulum
