+++ .github/workflows/BipedRobot.yml
@@ -13,7 +13,7 @@ jobs:
     - uses: julia-actions/setup-julia@latest
       with:
         # Build documentation on the latest Julia 1.x
-        version: '1'
+        version: '1.10'
     - name: Install dependencies
       run: julia --project=BipedRobot/ -e 'using Pkg; Pkg.instantiate();
                                      Pkg.develop(PackageSpec(path=pwd()))'
