### 2024-10-31 6b5ca69
Move solver parameters to attributes (#378)
* Move solver parameters to attributes

* Fix format

* Fixes

* Fixes

* Fix format

* Fix format

* Add missing time step

* BipedRobot with Julia v1.10

* Fixes
