### 2024-10-31 d72a253
Revert "Delete double pendulum"
This reverts commit 6f36279a2685ca8d7cf565c39c8950b9a492bf1d.
