+++ dev/generated/DC-DC converter.jl
@@ -11,7 +11,7 @@ const AB = OP.Abstraction
 
 include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "dc_dc.jl"))
 
-concrete_problem = DCDC.problem(; approx_mode = DCDC.GROWTH)
+concrete_problem = DCDC.problem()
 concrete_system = concrete_problem.system
 
 x0 = SVector(0.0, 0.0)
@@ -26,6 +26,8 @@ optimizer = MOI.instantiate(AB.UniformGridAbstraction.Optimizer)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("concrete_problem"), concrete_problem)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("state_grid"), state_grid)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("input_grid"), input_grid)
+MOI.set(optimizer, MOI.RawOptimizerAttribute("jacobian_bound"), DCDC.jacobian_bound())
+MOI.set(optimizer, MOI.RawOptimizerAttribute("time_step"), 0.5)
 MOI.optimize!(optimizer)
 
 abstract_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_controller"))
@@ -33,8 +35,12 @@ concrete_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("concrete_con
 
 nstep = 300
 x0 = SVector(1.2, 5.6)
-control_trajectory =
-    ST.get_closed_loop_trajectory(concrete_system.f, concrete_controller, x0, nstep)
+control_trajectory = ST.get_closed_loop_trajectory(
+    MOI.get(optimizer, MOI.RawOptimizerAttribute("discretized_system")),
+    concrete_controller,
+    x0,
+    nstep,
+)
 
 fig = plot(; aspect_ratio = :equal);
 plot!(concrete_system.X);
@@ -42,7 +48,7 @@ plot!(control_trajectory)
 
 include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "dc_dc.jl"))
 
-concrete_problem = DCDC.problem(; approx_mode = DCDC.DELTA_GAS)
+concrete_problem = DCDC.problem()
 concrete_system = concrete_problem.system
 
 origin = SVector(0.0, 0.0)
@@ -60,7 +66,13 @@ optimizer = MOI.instantiate(AB.UniformGridAbstraction.Optimizer)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("concrete_problem"), concrete_problem)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("state_grid"), state_grid)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("input_grid"), input_grid)
+MOI.set(
+    optimizer,
+    MOI.RawOptimizerAttribute("approx_mode"),
+    Dionysos.Optim.Abstraction.UniformGridAbstraction.DELTA_GAS,
+)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("δGAS"), true)
+MOI.set(optimizer, MOI.RawOptimizerAttribute("time_step"), 0.5)
 MOI.optimize!(optimizer)
 
 abstract_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_controller"))
@@ -68,8 +80,12 @@ concrete_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("concrete_con
 
 nstep = 300
 x0 = SVector(1.2, 5.6)
-control_trajectory =
-    ST.get_closed_loop_trajectory(concrete_system.f, concrete_controller, x0, nstep)
[…]
