### 2024-11-03 63723d7
modifiying uniformgrid abtraction to fit discrete time problems
