+++ docs/src/examples/solvers/Unicycle robot copy.jl
@@ -50,8 +50,8 @@ discretization_step = 0.1
 # Define the state variables: x1(t), x2(t), x3(t) for t = 1, ..., N
 x_low = [-3.5, -2.6, -pi]
 x_upp = -x_low
-x_initial = [0.0, 0.0, 0.0]  # Initial state
-#x_initial = [1.0, -1.7, 0.0]
+#x_initial = [0.0, 0.0, 0.0]  # Initial state
+x_initial = [1.0, -1.7, 0.0]
 @variable(model, x_low[i] <= x[i = 1:3] <= x_upp[i])#, start = x_initial[i])
 
 # Define the control variables: u1(t), u2(t) for t = 1, ..., N-1
@@ -66,8 +66,8 @@ x_initial = [0.0, 0.0, 0.0]  # Initial state
 
 # Define the initial and target sets
 
-x_target = [0.5, 0.5, -pi]  # Target state
-#x_target = [sqrt(32)/3, sqrt(20)/3, -pi]
+#x_target = [0.5, 0.5, -pi]  # Target state
+x_target = [sqrt(32)/3, sqrt(20)/3, -pi]
 
 ## constraint on the initial state+discretization_step
 #@constraint(model, start(x[1]) in MOI.Interval(x_initial[1] - discretization_step, x_initial[1] + discretization_step))
@@ -79,24 +79,28 @@ x_target = [0.5, 0.5, -pi]  # Target state
 # writing final(x[?, N]) in Interval is redundant. We should be able to write final(x[?]) in Interval
 
 #FIXME: if a julia variable is provided we should be able to handle it or to give more clear feedback
-#@constraint(model, final(x[1,N]) in MOI.Interval(x_target[1] - discretization_step, x_target[1] + discretization_step))
-#@constraint(model, final(x[2,N]) in MOI.Interval(x_target[2] - discretization_step, x_target[2] + discretization_step))
-#@constraint(model, final(x[3,N]) in MOI.Interval(x_target[3], x_target[3]))
+#@constraint(model, final(x[1]) in MOI.Interval(x_target[1] - discretization_step, x_target[1] + discretization_step))
+#@constraint(model, final(x[2]) in MOI.Interval(x_target[2] - discretization_step, x_target[2] + discretization_step))
+#@constraint(model, final(x[3]) in MOI.Interval(x_target[3], x_target[3]))
 
 #@constraint(model, start(x[1]) in MOI.Interval(x_initial[1] - discretization_step, x_initial[1] + discretization_step))
 #@constraint(model, start(x[2]) in MOI.Interval(x_initial[2] - discretization_step, x_initial[2] + discretization_step))
 #@constraint(model, start(x[3]) in MOI.Interval(x_initial[3], x_initial[3]))
 
-@constraint(model, start(x[1]) in MOI.Interval(-0.2, 0.2))
-@constraint(model, start(x[2]) in MOI.Interval(-0.2, 0.2))
-@constraint(model, start(x[3]) in MOI.Interval(-0.2, 0.2))
-
-#@constraint(model, start(x[1]) in MOI.Interval(0.8, 1.2))
-#@constraint(model, start(x[2]) in MOI.Interval(-1.9, -1.5))
+#@constraint(model, start(x[1]) in MOI.Interval(-0.2, 0.2))
+#@constraint(model, start(x[2]) in MOI.Interval(-0.2, 0.2))
 #@constraint(model, start(x[3]) in MOI.Interval(-0.2, 0.2))
 
-@constraint(model, final(x[1]) in MOI.Interval(0.3, 0.7))
[…]
