### 2024-11-03 fa48241
removing the copy file and the prints
