+++ docs/src/examples/solvers/Unicycle robot copy.jl
@@ -1,236 +1,0 @@
-using Test     #src
-# # Example: Unicycle Robot solved by [Uniform grid abstraction](https://github.com/dionysos-dev/Dionysos.jl/blob/master/docs/src/manual/manual.md#solvers).
-#
-# This example was adapted from the numerical experiments in [1, Sec. 5] from Symbolically guided Model Predictive Control (SgMPC)  [paper](https://doi.org/10.1016/j.ifacol.2022.09.039)..
-# This is a **control problem** for a **discrete-time nonlinear system**.
-#
-# Let us consider the 3-dimensional state space control system of the form
-# ```math
-# x_{t+1} = f(x_t, u_t)
-# ```
-# with $f: \mathbb{R}^3 × \mathbb{R}^2 \to \mathbb{R}^3$ given by
-# ```math
-# f(x, (u_1, u_2)) = \begin{bmatrix} x_1 + u_1 \cos(x_3) \\ x_2 + u_1 \sin(x_3) \\ x_3 + u_2 \ (\text{mod} \ 2\pi) \end{bmatrix}
-# ```
-# and with state and control constraints given by:
-# ```math
-# X = \left\{ (x_1, x_2)^T \in \mathbb{R}^2 \ | \ x_1^2 - x_2^2 \leq 4, \ 4x_2^2 - x_1^2 \leq 16 \right\}
-# ```
-# ```math
-# U = [0.2, 2] \times [-1, 1]
-# ```
-# Here, $(x_1, x_2)$ represents the 2D Cartesian coordinates and $x_3$ is the angular orientation of a mobile cart.
-# The control inputs $u_1$ and $u_2$ are the linear and angular velocities.
-# The control objective is to drive the mobile cart to a desired reference position $x_r$.
-#
-# Considering this as a reachability problem, we will use it to showcase the capabilities of the Uniform grid abstraction solving discrete-time problem in Dionysos.
-# The nonlinear constraints are handled as obstacles in the state-space.
-
-# First, let us import [StaticArrays](https://github.com/JuliaArrays/StaticArrays.jl) and [Plots](https://github.com/JuliaPlots/Plots.jl).
-using StaticArrays, Plots, Revise
-
-# At this point, we import Dionysos and JuMP.
-using Dionysos, JuMP
-
-# Define the problem using JuMP
-# We first create a JuMP model:
-model = Model(Dionysos.Optimizer)
-
-# Define the discretization step
-discretization_step = 0.1
-
-# Define the state variables: x1(t), x2(t), x3(t) without specifying the start value since here it's a set. We will specify the start later using constraints.
-x_low = [-3.5, -2.6, -pi]
-x_upp = -x_low
-@variable(model, x_low[i] <= x[i = 1:3] <= x_upp[i])
-
-# Define the control variables: u1(t), u2(t)
-@variable(model, -1 <= u[1:2] <= 1)
-
-# Define the dynamics
-@constraint(model, ∂(x[1]) == x[1] + u[1] * cos(x[3]))
-@constraint(model, ∂(x[2]) == x[2] + u[1] * sin(x[3]))
-@constraint(model, ∂(x[3]) == x[3] + u[2])
-
-# Define the initial and target sets
-x_initial = [1.0, -1.7, 0.0]
-x_target = [sqrt(32)/3, sqrt(20)/3, -pi]
-
-@constraint(model, start(x[1]) in MOI.Interval(0.8, 1.2))
-@constraint(model, start(x[2]) in MOI.Interval(-1.9, -1.5))
-@constraint(model, start(x[3]) in MOI.Interval(-0.2, 0.2))
-
-@constraint(model, final(x[1]) in MOI.Interval(1.78, 1.98))
-@constraint(model, final(x[2]) in MOI.Interval(1.39, 1.59))
[…]
