### 2024-11-03 45575ec
updating the wrapper to allow adding initial values when defining the variables using the start parameter
