### 2024-11-05 2196646
DSICRETE -> DISCRETE
