+++ src/optim/abstraction/uniform_grid_abstraction.jl
@@ -81,6 +81,62 @@ function MOI.get(model::Optimizer, param::MOI.RawOptimizerAttribute)
     return getproperty(model, Symbol(param.name))
 end
 
+function discretized_system(concrete_system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem, model, noise)
+    if isnothing(model.growthbound_map)
+        error("Please set the `growthbound_map`.")
+    end
+    if isnothing(model.sys_inv_map)
+        error("Please set the `sys_inv_map`.")
+    end
+    return Dionysos.System.ControlSystemGrowth(
+        1.0, # `time_step` should be ignored by `concrete_system.f`, `model.growthbound_map` and `model.sys_inv_map` anyway
+        noise,
+        noise,
+        concrete_system.f,
+        model.growthbound_map,
+        model.sys_inv_map,
+    )
+end
+
+function discretized_system(concrete_system::MathematicalSystems.ConstrainedBlackBoxControlContinuousSystem, model, noise)
+    if isnan(model.time_step)
+        error("Please set the `time_step`.")
+    end
+
+    if model.approx_mode == GROWTH
+        if isnothing(model.jacobian_bound)
+            error("Please set the `jacobian_bound`.")
+        end
+        return Dionysos.System.NewControlSystemGrowthRK4(
+            model.time_step,
+            concrete_system.f,
+            model.jacobian_bound,
+            noise,
+            noise,
+            model.num_sub_steps_system_map,
+            model.num_sub_steps_growth_bound,
+        )
+    elseif model.approx_mode == LINEARIZED
+        return Dionysos.System.NewControlSystemLinearizedRK4(
+            model.time_step,
+            concrete_system.f,
+            model.jacobian,
+            u -> 0.0,
+            u -> 0.0,
+            noise,
+            model.num_sub_steps_system_map,
+        )
+    else
+        @assert model.approx_mode == DELTA_GAS
+        return Dionysos.System.NewSimpleSystem(
+            model.time_step,
+            concrete_system.f,
+            noise,
+            model.num_sub_steps_system_map,
+        )
+    end
+end
+
 function build_abstraction(concrete_system, model)
     Xfull = DO.DomainList(model.state_grid)
     DO.add_set!(Xfull, concrete_system.X, DO.INNER)
@@ -93,58 +149,7 @@ function build_abstraction(concrete_system, model)
     nstates = Dionysos.Utils.get_dims(concrete_system.X)
     noise = StaticArrays.SVector(ntuple(_ -> 0.0, Val(nstates)))
 
-    if isnan(model.time_step)
-        error("Please set the `time_step`.")
-    end
-
-    model.discretized_system =
-        if concrete_system isa MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem
-            if isnothing(model.growthbound_map)
-                error("Please set the `growthbound_map`.")
-            end
-            if isnothing(model.sys_inv_map)
-                error("Please set the `sys_inv_map`.")
-            end
-            Dionysos.System.ControlSystemGrowth(
-                model.time_step,
-                noise,
-                noise,
-                concrete_system.f,
-                model.growthbound_map,
-                model.sys_inv_map,
-            )
-        elseif model.approx_mode == GROWTH
-            if isnothing(model.jacobian_bound)
-                error("Please set the `jacobian_bound`.")
-            end
-            Dionysos.System.NewControlSystemGrowthRK4(
-                model.time_step,
-                concrete_system.f,
-                model.jacobian_bound,
-                noise,
-                noise,
-                model.num_sub_steps_system_map,
-                model.num_sub_steps_growth_bound,
-            )
-        elseif model.approx_mode == LINEARIZED
-            Dionysos.System.NewControlSystemLinearizedRK4(
[…]
