### 2024-11-06 63d669d
Adding unicycle in the new frame (#383)
* adding unicycle in the new frame

* updating the wrapper

* jformat

* modifiying uniformgrid abtraction to fit discrete time problems

* cleaning

* Adding docs comments

* removing the copy file and the prints

* format

* updating the wrapper to allow adding initial values when defining the variables using the start parameter

* format

* removing revize

* complying with litterate

* linking to the discretization step

* DSICRETE -> DISCRETE

* Use Delta (#386)

* Fix format

* Checkout JuMP master

* Remove time_step

* Fix format

* Fix comment


[…]
