### 2024-11-07 07d3741
capturing general non linearity in case of var in an interval
