### 2024-11-08 8cc8acb
Remove modulo in sys_inv_map
