### 2024-11-08 0c2c86f
Rename discretization functions
