+++ docs/src/examples/Getting Started.jl
@@ -80,7 +80,7 @@ measnoise = SVector(0.0, 0.0);
 sysnoise = SVector(0.0, 0.0);
 
 # And now the instantiation of the ControlSystem
-contsys = ST.NewControlSystemGrowthRK4(
+contsys = ST.discretize_system_with_growth_bound(
     tstep,
     F_sys,
     L_growthbound,
