### 2024-11-08 97061cb
Refactor towards MathematicalSystems.jl
