+++ src/system/controlsystem.jl
@@ -6,14 +6,11 @@ The structure
 is the abstract type that defines a control system.
 """
 abstract type ControlSystem{N, T} end
-using IntervalArithmetic
-function get_f_eval(sys)
-    return sys.f_eval
-end
+import IntervalArithmetic
 
-function RungeKutta4(F, x, u, tstep, nsub::Int)
-    τ = tstep / nsub
-    for _ in 1:nsub
+function runge_kutta_4(F, x, u, Δt, num_sub_steps::Int)
+    τ = Δt / num_sub_steps
+    for _ in 1:num_sub_steps
         Fx1 = F(x, u)
         xrk = x + Fx1 * (τ / 2.0)
         Fx2 = F(xrk, u)
@@ -26,51 +23,59 @@ function RungeKutta4(F, x, u, tstep, nsub::Int)
     return x
 end
 
-"""
-    ControlSystemGrowth{N, T, F1 <: Function, F2 <: Function, F3 <: Function} <: ControlSystem{N, T}
-
-is one implementation of the `ControlSystem` type for which we have a growth bound function.
-"""
-struct ControlSystemGrowth{N, T, F1 <: Function, F2 <: Function, F3 <: Function} <:
-       ControlSystem{N, T}
-    tstep::Float64
-    sysnoise::SVector{N, T}
-    measnoise::SVector{N, T}
-    sys_map::F1
-    growthbound_map::F2
-    sys_inv_map::F3
+struct RungeKuttaDiscretization <: MathematicalSystems.AbstractDiscretizationAlgorithm
+    num_sub_steps::Int
 end
 
-function discretize_system_with_growth_bound(
-    tstep,
-    F_sys,
-    L_growthbound,
-    sysnoise::SVector{N, T},
-    measnoise::SVector{N, T},
-    nsys,
-    ngrowthbound,
-) where {N, T}
-    sys_map = let nsys = nsys
-        (x::SVector{N, T}, u, tstep) -> RungeKutta4(F_sys, x, u, tstep, nsys)::SVector{N, T}
+function MathematicalSystems.discretize(
+    system::MathematicalSystems.BlackBoxControlContinuousSystem,
+    Δt,
+    algorithm::RungeKuttaDiscretization,
+)
+    f = let
+        num_sub_steps = algorithm.num_sub_steps
+        Δt = Δt
+        (x::SVector{N, T}, u) ->
+            runge_kutta_4(system.f, x, u, Δt, num_sub_steps)::SVector{N, T}
     end
-    sys_inv_map = let nsys = nsys
-        (x::SVector{N, T}, u, tstep) ->
-            RungeKutta4(F_sys, x, u, -tstep, nsys)::SVector{N, T}
+    return MathematicalSystems.BlackBoxControlDiscreteSystem(
+        f,
+        system.statedim,
+        system.inputdim,
+    )
+end
+
+"""
+    struct WithGrowthBound{S <: MathematicalSystems.AbstractSystem, F <: Function} <: MathematicalSystems.AbstractSystem
+        system::S
+        growth_bound::F
     end
-    F_growthbound = let sysnoise = sysnoise
-        (r, u) -> L_growthbound(u) * r + sysnoise
+
+System with growth bound function.
+"""
+struct WithGrowthBound{S <: MathematicalSystems.AbstractSystem, F <: Function} <: MathematicalSystems.AbstractSystem
+    system::S
+    growth_bound::F
+end
+
+function MathematicalSystems.discretize(
+    system::WithGrowthBound,
+    Δt,
+    algorithm::RungeKuttaDiscretization,
+)
+    growth_bound = let
[…]
