### 2024-11-08 8dc58bc
Rename discretization functions (#388)
