+++ dev/generated/Unicycle robot.jl
@@ -95,9 +95,9 @@ set_attribute(model, "growthbound_map", growth_bound)
 
 function sys_inv(x, u, _)
     return StaticArrays.SVector{3}(
-        x[1] - u[1] * cos((x[3] - u[2]) % (2 * π)),
-        x[2] - u[1] * sin((x[3] - u[2]) % (2 * π)),
-        (x[3] - u[2]) % (2 * π),
+        x[1] - u[1] * cos(x[3] - u[2]),
+        x[2] - u[1] * sin(x[3] - u[2]),
+        x[3] - u[2],
     )
 end
 set_attribute(model, "sys_inv_map", sys_inv)
