### 2024-11-08 9ac4f50
Remove SymPy
