+++ docs/make.jl
@@ -75,7 +75,4 @@ makedocs(;
     ],
 )
 
-deploydocs(;
-    repo = "github.com/dionysos-dev/Dionysos.jl.git",
-    push_preview = true,
-)
+deploydocs(; repo = "github.com/dionysos-dev/Dionysos.jl.git", push_preview = true)
