+++ previews/PR390/examples/solvers/Gol, Lazar & Belta (2013).jl
@@ -1,0 +1,146 @@
+using Test     #src
+# # Example: Gol, Lazar and Belta (2013) solved by [Bemporad Morari](https://github.com/dionysos-dev/Dionysos.jl/blob/master/docs/src/manual/manual.md#solvers).
+#
+#md # [![Binder](https://mybinder.org/badge_logo.svg)](@__BINDER_ROOT_URL__/generated/Gol%2C Lazar %26 Belta (2013).ipynb)
+#md # [![nbviewer](https://img.shields.io/badge/show-nbviewer-579ACA.svg)](@__NBVIEWER_ROOT_URL__/generated/Gol%2C Lazar %26 Belta (2013).ipynb)
+# This example reproduces parts of the numerical results of [3]. A similar example reproducing all results of [3] is available as a codeocean capsule in [4].
+#
+# This example was borrowed from [1, Example VIII.A] and tackles
+# an optimal control for the hybrid system with state evolution governed by
+# ```math
+# x(k+1) = \begin{bmatrix} 1 & 1 \\ 0 & 1 \end{bmatrix}x(k) + \begin{bmatrix} 0.5 \\ 1.0 \end{bmatrix} u(k)
+# ```
+
+# The goal is to take the state vector toward a target set **XT** by visiting one of the squares
+# **A** or **B** and avoiding the obstacles **O1** and **O2**
+
+# First, let us import [CDDLib](https://github.com/JuliaPolyhedra/CDDLib.jl),
+# [GLPK](https://github.com/jump-dev/GLPK.jl), [OSQP](https://github.com/oxfordcontrol/OSQP.jl),
+# [JuMP](https://github.com/jump-dev/JuMP.jl), [Pavito](https://github.com/jump-dev/Pavito.jl)
+# and [Ipopt](https://github.com/jump-dev/Ipopt.jl)
+import CDDLib
+import GLPK
+import OSQP
+using JuMP
+import Pavito
+import HiGHS
+import Ipopt
+
+# At this point we import Dionysos
+using Dionysos
+const DI = Dionysos
+const UT = DI.Utils
+const ST = DI.System
+const OP = DI.Optim
+
+# And the file defining the hybrid system for this problem
+include(joinpath(dirname(dirname(pathof(Dionysos))), "problems", "gol_lazar_belta.jl"))
+
+# Now we instantiate our optimal control problem using the function provided by [GolLazarBelta.jl](@__REPO_ROOT_URL__/problems/GolLazarBelta.jl)
+problem = GolLazarBelta.problem(CDDLib.Library(), Float64);
+
+# Finally, we select the method presented in [2] as our optimizer
+
+qp_solver = optimizer_with_attributes(
+    OSQP.Optimizer,
+    "eps_abs" => 1e-8,
+    "eps_rel" => 1e-8,
+    "max_iter" => 100000,
+    MOI.Silent() => true,
+);
+
+mip_solver = optimizer_with_attributes(HiGHS.Optimizer, MOI.Silent() => true);
+
+cont_solver = optimizer_with_attributes(Ipopt.Optimizer, MOI.Silent() => true);
+
+miqp_solver = optimizer_with_attributes(
+    Pavito.Optimizer,
+    "mip_solver" => mip_solver,
+    "cont_solver" => cont_solver,
+    MOI.Silent() => true,
+);
+
+algo = optimizer_with_attributes(
+    OP.BemporadMorari.Optimizer{Float64},
+    "continuous_solver" => qp_solver,
+    "mixed_integer_solver" => miqp_solver,
+    "indicator" => false,
+    "log_level" => 0,
+);
+
+# and use it to solve the given problem, with the help of the abstraction layer
[…]
