### 2024-11-08 e885a31
BipedRobot back on Julia v1.11 (#391)
* BipedRobot back on Julia v1.11

* Remove SymPy
