### 2024-11-08 33bfbf6
Biblio with DocumenterBibliography (#390)
* Biblio with DocumenterBibliography

* Remove extra line

* Use const

* Fix

* Fix

* Fix

* Fix format
