+++ docs/Project.toml
@@ -4,6 +4,7 @@ Clarabel = "61c947e1-3e6d-4ee4-985a-eec8c727bd6e"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 Dionysos = "d92c97cf-b87d-42c1-a9c0-25df00b4d958"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
@@ -33,4 +34,5 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 Documenter = "1"
+DocumenterCitations = "1"
 Libtiff_jll = "4.4.0"
