+++ Project.toml
@@ -25,12 +25,14 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
 MathOptSymbolicAD = "309f4015-3481-4d63-a8f9-aeb13adfe8eb"
 MathematicalSystems = "d14a8603-c872-5ed3-9ece-53e0e82e39da"
+MeshCatMechanisms = "6ad125db-dd91-5488-b820-c1df6aab299d"
 MutableArithmetics = "d8a4904e-b15c-11e9-3269-09a3773c0cb0"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+RigidBodyDynamics = "366cf18f-59d5-5db9-a4de-86a9f6786172"
 SemialgebraicSets = "8e049039-38e8-557d-ae3a-bc521ccf6204"
 SimpleWeightedGraphs = "47aef6b3-ad0c-573a-a1e2-d07658019622"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
@@ -60,10 +62,12 @@ LightXML = "0.9"
 MathOptInterface = "1"
 MathOptSymbolicAD = "0.2.1"
 MathematicalSystems = "0.11, 0.12, 0.13"
+MeshCatMechanisms = "0.9.0"
 MutableArithmetics = "1"
 Plots = "1"
 Polyhedra = "0.7"
 ProgressMeter = "1.4"
+RigidBodyDynamics = "2.5.0"
 SemialgebraicSets = "0.2, 0.3"
 SimpleWeightedGraphs = "1.2"
 SpecialFunctions = "2"
