+++ docs/src/examples/solvers/Pendulum_Safety.jl
@@ -29,7 +29,11 @@ MOI.set(optimizer, MOI.RawOptimizerAttribute("concrete_problem"), concrete_probl
 MOI.set(optimizer, MOI.RawOptimizerAttribute("state_grid"), state_grid)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("input_grid"), input_grid)
 MOI.set(optimizer, MOI.RawOptimizerAttribute("time_step"), 0.3)
-MOI.set(optimizer, MOI.RawOptimizerAttribute("jacobian_bound"), concrete_system.f.growthbound_map)
+MOI.set(
+    optimizer,
+    MOI.RawOptimizerAttribute("jacobian_bound"),
+    concrete_system.f.growthbound_map,
+)
 MOI.optimize!(optimizer)
 
 abstract_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("abstract_controller"))
