### 2024-11-11 2f65e11
Adding the single pendulum example (#392)
* Adding the single pendulum example

* Formatting

* Merge branch master
