+++ previews/PR394/examples/solvers/single pendulum reach.jl
@@ -67,6 +67,7 @@ x1_target, x2_target = pi .+ (5.0 * pi / 180.0) .* [-1, 1], 1.0 .* [-1, 1]
 
 @constraint(model, final(x[1]) in MOI.Interval(x1_target...))
 @constraint(model, final(x[2]) in MOI.Interval(x2_target...))
+nothing #hide
 
 # ### Definition of the abstraction
 
