### 2024-11-12 d6bcbe7
Cosmetic changes to single pendulum (#394)
* Cosmetic changes to single pendulum

* Add newline

* adding one more nothing-hide

---------

Co-authored-by: johnaoga <johnaoga@gmail.com>
