### 2024-11-13 1353d5d
adding guard and mode
