+++ src/MOI_wrapper.jl
@@ -28,8 +28,14 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
     objective_function::MOI.AbstractScalarFunction
     nonlinear_index::Int
     modes_table::Vector{Union{Nothing, Tuple{MOI.VariableIndex, Function}}} # mode_index -> (mode_var, mode_func)
-    guards_table::Vector{Union{Nothing, Tuple{MOI.VariableIndex, MOI.VariableIndex , Function}}}  # guard_index -> (mode_var_src, mode_var_dst, guard_func)
-    resetmaps_table::Vector{Union{Nothing, Tuple{Function , Function}}}  # resetmap_index -> (mode_func, resetmap_func)
+    guards_table::Vector{
+        Union{Nothing, Tuple{MOI.VariableIndex, MOI.VariableIndex, Function}},
+    }  # guard_index -> (mode_var_src, mode_var_dst, guard_func)
+    resetmaps_table::Vector{Union{Nothing, Tuple{Function, Function}}}  # resetmap_index -> (mode_func, resetmap_func)
+    integer_variables::Dict{MOI.VariableIndex, Union{MOI.Integer, MOI.ZeroOne}}
+    indicator_constraints::Vector{
+        Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction},
+    }
     function Optimizer()
         return new(
             MOI.instantiate(Dionysos.Optim.Abstraction.UniformGridAbstraction.Optimizer),
@@ -50,8 +56,10 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
             zero(MOI.ScalarAffineFunction{Float64}),
             0,
             Union{Nothing, Tuple{MOI.VariableIndex, Function}}[],
-            Union{Nothing, Tuple{MOI.VariableIndex, MOI.VariableIndex , Function}}[],
-            Union{Nothing, Tuple{Function , Function}}[],
+            Union{Nothing, Tuple{MOI.VariableIndex, MOI.VariableIndex, Function}}[],
+            Union{Nothing, Tuple{Function, Function}}[],
+            Dict{MOI.VariableIndex, Union{MOI.Integer, MOI.ZeroOne}}(),
+            Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction}[],
         )
     end
 end
@@ -76,6 +84,8 @@ function MOI.empty!(model::Optimizer)
     empty!(model.modes_table)
     empty!(model.guards_table)
     empty!(model.resetmaps_table)
+    empty!(model.integer_variables)
+    empty!(model.indicator_constraints)
     return
 end
 
@@ -251,6 +261,37 @@ function MOI.set(
     return
 end
 
+# Adding supports for integer variables
+function MOI.supports_constraint(
+    ::Optimizer,
+    ::Type{MOI.VariableIndex},
+    ::Type{<:Union{MOI.Integer, MOI.ZeroOne}},
+)
+    return true
+end
+
+function MOI.add_constraint(
+    model::Optimizer,
+    func::MOI.VariableIndex,
+    set::Union{MOI.Integer, MOI.ZeroOne},
+)
+    model.integer_variables[func] = set
+    return MOI.ConstraintIndex{typeof(func), typeof(set)}(func.value)
+end
+
+# Adding support for indicator constraints
+function JuMP._build_indicator_constraint(
+    error_fn::Function,
+    lhs::JuMP.NonlinearExpr,
+    constraint::JuMP.ScalarConstraint,
+    ::Type{MOI.Indicator{A}},
+) where {A}
+    #return error("Unsupported")
+    set = MOI.Indicator{A}(JuMP.moi_set(constraint))
[…]
