### 2024-11-13 ee1365e
adding thermostat example
