### 2024-11-16 e2d7a4b
removing guard function
