### 2024-11-16 f5bba76
adding transaction struct
