### 2024-11-18 f991ba6
adding support for constraints with transactions
