### 2024-11-18 3d5b841
reworking the examples
