### 2024-11-18 008aaa6
adding the syntactic sugar ;-)
