+++ docs/src/examples/solvers/thermostat.jl
@@ -62,14 +62,31 @@ T_start = 10.0
 @constraint(model, mode == 1 => {∂(T) == k * u})
 @constraint(model, mode == 0 => {∂(T) == -α})
 
+# Define the guards and resetmaps for each transition
+
+## Transition 0 -> 1
+#@constraint(model, guard(mode == 1, mode == 0) => {T >= T_r - δ})
+#@constraint(model, resetmaps(mode == 0, mode == 1) => {Δ(T) == ...})
+add_transition!(model, mode, 0, 1) do t
+    @constraint(t, T <= T_r + δ) # guard
+    #@constraint(t, Δ(T) == ...) # resetmap
+end
+
+## Transition 1 -> 0
+#@constraint(model, guard(mode == 0, mode == 1) => {T <= T_r + δ})
+#@constraint(model, resetmaps(mode == 1, mode == 0) => {Δ(T) == ...})
+add_transition!(model, mode, 1, 0) do t
+    @constraint(t, T >= T_r - δ) # guard
+    #@constraint(t_10, Δ(T) == ...) # resetmap
+end
 
 # Define the guards
 #@constraint(model, guard(mode == 1, mode == 0) => {T >= T_r - δ})
 #@constraint(model, guard(mode == 0, mode == 1) => {T <= T_r + δ})
-t_01 = Transition(model, mode, 0, 1)
-@constraint(t_01, T <= T_r + δ)
-t_10 = Transition(model, mode, 1, 0)
-@constraint(t_10, T >= T_r - δ)
+##t_01 = Transition(model, mode, 0, 1)
+##@constraint(t_01, T <= T_r + δ)
+##t_10 = Transition(model, mode, 1, 0)
+##@constraint(t_10, T >= T_r - δ)
 
 # Define the resetmaps
 #@constraint(model, resetmaps(mode == 0, mode == 1) => {Δ(T) == ...})
