### 2024-11-19 3efdfae
get rid of vector constraint, store modes, guards, and resetmaps
