+++ src/MOI_wrapper.jl
@@ -5,7 +5,7 @@ import JuMP
 import MathOptSymbolicAD
 import Symbolics
 
-export ∂, Δ, final, start, rem, Transition, add_transition!
+export ∂, Δ, final, start, rem, Transition, add_transition!, guard, resetmaps, themode
 
 @enum(VariableType, INPUT, STATE, MODE)
 
@@ -29,11 +29,9 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
     objective_sense::MOI.OptimizationSense
     objective_function::MOI.AbstractScalarFunction
     nonlinear_index::Int
-    modes_table::Vector{Union{Nothing, Tuple{MOI.VariableIndex, Function}}} # mode_index -> (mode_var, mode_func)
-    guards_table::Vector{
-        Union{Nothing, Tuple{MOI.VariableIndex, MOI.VariableIndex, Function}},
-    }  # guard_index -> (mode_var_src, mode_var_dst, guard_func)
-    resetmaps_table::Vector{Union{Nothing, Tuple{Function, Function}}}  # resetmap_index -> (mode_func, resetmap_func)
+    modes_table::Vector{Union{Nothing, Tuple{JuMP.VariableRef, Int64, Function}}} # mode_index -> (mode_var, mode_val, mode_func)
+    guards_table::Vector{Union{Nothing, Tuple{JuMP.VariableRef, Int64, Int64, Function}}}  # guard_index -> (mode_var, src, dst, guard_func)
+    resetmaps_table::Vector{Union{Nothing, Tuple{JuMP.VariableRef, Int64, Int64, Function}}}  # resetmap_index -> (mode_var, src, dst, resetmap_func)
     integer_variables::Dict{MOI.VariableIndex, Union{MOI.Integer, MOI.ZeroOne}}
     indicator_constraints::Vector{
         Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction},
@@ -57,9 +55,9 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
             MOI.FEASIBILITY_SENSE,
             zero(MOI.ScalarAffineFunction{Float64}),
             0,
-            Union{Nothing, Tuple{MOI.VariableIndex, Function}}[],
-            Union{Nothing, Tuple{MOI.VariableIndex, MOI.VariableIndex, Function}}[],
-            Union{Nothing, Tuple{Function, Function}}[],
+            Union{Nothing, Tuple{JuMP.VariableRef, Int64, Function}}[],
+            Union{Nothing, Tuple{JuMP.VariableRef, Int64, Int64, Function}}[],
+            Union{Nothing, Tuple{JuMP.VariableRef, Int64, Int64, Function}}[],
             Dict{MOI.VariableIndex, Union{MOI.Integer, MOI.ZeroOne}}(),
             Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction}[],
         )
@@ -288,9 +286,75 @@ function JuMP._build_indicator_constraint(
     constraint::JuMP.ScalarConstraint,
     ::Type{MOI.Indicator{A}},
 ) where {A}
-    #return error("Unsupported")
-    set = MOI.Indicator{A}(JuMP.moi_set(constraint))
-    return JuMP.VectorConstraint([lhs, JuMP.jump_function(constraint)], set)
+    # Get the function and set
+    fun, set = JuMP.jump_function(constraint), JuMP.moi_set(constraint)
+    model = JuMP.owner_model(lhs)
+
+    # Check if the function is a aff expression in which case convert it into a nonlinear expression
+    stored_fun = fun
+    if fun isa JuMP.AffExpr
[…]
