### 2024-11-19 5ad086e
fitting original variable philosophy
