+++ src/MOI_wrapper.jl
@@ -280,6 +280,72 @@ function MOI.add_constraint(
 end
 
 # Adding support for indicator constraints
+function _handle_affExpr(fun, set)
+    if !(fun isa JuMP.AffExpr)
+        error("Unsupported function type. `AffExpr` expected.")
+    end
+
+    if set isa MOI.LessThan
+        return JuMP.NonlinearExpr(:(<=), fun, set.upper)
+    elseif set isa MOI.GreaterThan
+        return JuMP.NonlinearExpr(:(>=), fun, set.lower)
+    elseif set isa MOI.EqualTo
+        return JuMP.NonlinearExpr(:(==), fun, set.value)
+    end
+
+    return error(
+        "Unsupported set type {$type(set)}. `LessThan`, `GreaterThan` or `EqualTo` expected.",
+    )
+end
+
+function _handle_hybrid_systems_table(model, expr, fun)
+    # Set hybrid systems tables
+
+    if !(length(expr.args) == 2)
+        error(
+            "Unsupported! If you are defining a mode the left-hand side must be something like mode == val. If you are defining a guard or resetmap, use add_transition.",
+        )
+    end
+
+    # handling mode
+    if expr.head == :(==)
+        lhs_var, lhs_val = expr.args[1], expr.args[2]
+        if !(lhs_var isa JuMP.VariableRef && lhs_val isa Int)
+            error(
+                "Unsupported! If you are defining a mode the left-hand side must be something like mode == val.",
+            )
+        end
+
+        if !haskey(model.ext, :modes)
+            model.ext[:modes] = []
+        end
+
+        push!(model.ext[:modes], (lhs_var, lhs_val, fun))
+    end
+
+    # handling guard
+    if expr.head == :guard
+        lhs_var, lhs_src, lhs_dst = expr.args[1].args[1], expr.args[1].args[2], expr.args[2]
+
+        if !haskey(model.ext, :guards)
+            model.ext[:guards] = []
+        end
+
+        push!(model.ext[:guards], (lhs_var, lhs_src, lhs_dst, fun))
+    end
+
+    # handling resetmap
+    if expr.head == :resetmaps
+        lhs_var, lhs_src, lhs_dst = expr.args[1].args[1], expr.args[1].args[2], expr.args[2]
+
+        if !haskey(model.ext, :resetmaps)
+            model.ext[:resetmaps] = []
+        end
+
+        push!(model.ext[:resetmaps], (lhs_var, lhs_src, lhs_dst, fun))
+    end
+end
+
 function JuMP._build_indicator_constraint(
     error_fn::Function,
     lhs::JuMP.NonlinearExpr,
@@ -289,68 +355,32 @@ function JuMP._build_indicator_constraint(
     # Get the function and set
     fun, set = JuMP.jump_function(constraint), JuMP.moi_set(constraint)
     model = JuMP.owner_model(lhs)
+    @show fun, set
+
+    if lhs isa JuMP.NonlinearExpr && fun isa JuMP.NonlinearExpr
+        return JuMP.build_constraint(
+            error_fn,
+            JuMP.NonlinearExpr(:(==), lhs, fun),
+            MOI.EqualTo(0.0),
+        )
+    end
+
+    if lhs isa JuMP.NonlinearExpr && fun isa JuMP.AffExpr
+        rhs = _handle_affExpr(fun, set)
+        return JuMP.build_constraint(
+            error_fn,
+            JuMP.NonlinearExpr(:(==), lhs, rhs),
+            MOI.EqualTo(0.0),
+        )
+    end
 
     # Check if the function is a aff expression in which case convert it into a nonlinear expression
-    stored_fun = fun
-    if fun isa JuMP.AffExpr
-        fun, set = constraint.func, constraint.set
[…]
