+++ src/MOI_wrapper.jl
@@ -187,6 +187,7 @@ function MOI.add_constraint(
     func::MOI.ScalarNonlinearFunction,
     set::MOI.EqualTo,
 )
+
     if func.head == :- && length(func.args) == 2
         lhs, rhs = func.args
         if lhs isa MOI.ScalarNonlinearFunction && length(lhs.args) == 1
@@ -230,21 +231,57 @@ function MOI.add_constraint(
     return error("Unsupported")
 end
 
-function MOI.supports(::Optimizer, ::Union{MOI.ObjectiveSense, MOI.ObjectiveFunction})
-    return true
-end
+function _handle_hybrid_systems_table(model, lhs, rhs)
+    # Set hybrid systems tables
 
-function MOI.supports(::Optimizer, ::MOI.VariablePrimalStart, ::Type{MOI.VariableIndex})
-    return true
-end
-
-function MOI.set(model::Optimizer, ::MOI.VariablePrimalStart, vi::MOI.VariableIndex, value)
-    # create a MOI.Interval from the value if value is a scalar
-    if !isa(value, MOI.Interval)
-        value = MOI.Interval(value, value)
+    if !(length(lhs.args) == 2)
+        error(
+            "Unsupported! If you are defining a mode the left-hand side must be something like mode == val. If you are defining a guard or resetmap, use add_transition.",
+        )
     end
 
-    return model.start[vi.value] = value
+    # handling mode
+    if lhs.head == :(==)
+        lhs_var, lhs_val = lhs.args[1], lhs.args[2]
+        if !(lhs_var isa JuMP.VariableRef && lhs_val isa Int)
+            error(
+                "Unsupported! If you are defining a mode the left-hand side must be something like mode == val.",
+            )
+        end
+
+        if !haskey(model.ext, :modes)
+            model.ext[:modes] = []
+        end
+
+        push!(model.ext[:modes], (lhs_var, lhs_val, rhs))
+    end
+
+    # handling guard
+    if lhs.head == :guard
+        lhs_var, lhs_src, lhs_dst = lhs.args[1].args[1], lhs.args[1].args[2], lhs.args[2]
+
+        if !haskey(model.ext, :guards)
+            model.ext[:guards] = []
+        end
+
+        push!(model.ext[:guards], (lhs_var, lhs_src, lhs_dst, rhs))
+    end
+
+    # handling resetmap
+    if lhs.head == :resetmaps
+        lhs_var, lhs_src, lhs_dst = lhs.args[1].args[1], lhs.args[1].args[2], lhs.args[2]
+
+        if !haskey(model.ext, :resetmaps)
+            model.ext[:resetmaps] = []
+        end
+
+        push!(model.ext[:resetmaps], (lhs_var, lhs_src, lhs_dst, rhs))
+    end
+end
+
+# Adding support for objective sense
+function MOI.supports(::Optimizer, ::Union{MOI.ObjectiveSense, MOI.ObjectiveFunction})
+    return true
 end
 
 function MOI.set(model::Optimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSense)
@@ -259,6 +296,20 @@ function MOI.set(
 )
     model.objective_function = func
     return
+end
+
+# Adding support for primal start
+function MOI.supports(::Optimizer, ::MOI.VariablePrimalStart, ::Type{MOI.VariableIndex})
+    return true
+end
+
+function MOI.set(model::Optimizer, ::MOI.VariablePrimalStart, vi::MOI.VariableIndex, value)
+    # create a MOI.Interval from the value if value is a scalar
+    if !isa(value, MOI.Interval)
+        value = MOI.Interval(value, value)
+    end
+
[…]
