### 2024-11-19 73e447b
handling guards and reset maps and cleaning
