+++ docs/src/examples/solvers/thermostat.jl
@@ -64,33 +64,24 @@ T_start = 10.0
 
 # Define the guards and resetmaps for each transition
 
-## Transition 0 -> 1
-#@constraint(model, guard(mode == 1, mode == 0) => {T >= T_r - δ})
-#@constraint(model, resetmaps(mode == 0, mode == 1) => {Δ(T) == ...})
-add_transition!(model, mode, 0, 1) do t
-    @constraint(t, T <= T_r + δ) # guard
-    @constraint(t, Δ(T) == k * u) # resetmap
-end
+# Transition 0 -> 1
+@constraint(model, guard(mode == 1, mode == 0) => {T >= T_r - δ})
 
-## Transition 1 -> 0
-#@constraint(model, guard(mode == 0, mode == 1) => {T <= T_r + δ})
-#@constraint(model, resetmaps(mode == 1, mode == 0) => {Δ(T) == ...})
-add_transition!(model, mode, 1, 0) do t
-    @constraint(t, T >= T_r - δ) # guard
-    #@constraint(t_10, Δ(T) == ...) # resetmap
-end
+# The resetmaps can be defined as follows:
+# ```julia 
+# @constraint(model, resetmaps(mode == 0, mode == 1) => {Δ(T) == ...})
+# ```
 
-# Define the guards
-#@constraint(model, guard(mode == 1, mode == 0) => {T >= T_r - δ})
-#@constraint(model, guard(mode == 0, mode == 1) => {T <= T_r + δ})
-##t_01 = Transition(model, mode, 0, 1)
-##@constraint(t_01, T <= T_r + δ)
-##t_10 = Transition(model, mode, 1, 0)
-##@constraint(t_10, T >= T_r - δ)
+# Alternatively we can define the transitions using `add_transition!` function
+# ```julia
+# add_transition!(model, mode, 0, 1) do t
+#    @constraint(t, T <= T_r + δ) # guard
+#    @constraint(t, Δ(T) == ...) # resetmap
+# end
+# ```
 
-# Define the resetmaps
-#@constraint(model, resetmaps(mode == 0, mode == 1) => {Δ(T) == ...})
-#@constraint(t_01, Δ(T) == ...)
+# Transition 1 -> 0
+@constraint(model, guard(mode == 0, mode == 1) => {T <= T_r + δ})
 
 # Define the initial set 
 @constraint(model, start(T) in MOI.Interval(9.5, 10.5))
@@ -113,5 +104,6 @@ u0 = SVector(0.0);
 h = SVector(0.3);
 set_attribute(model, "input_grid", Dionysos.Domain.GridFree(u0, h))
 
-# Solving the problem
+# ### Solving the problem
 optimize!(model)
+
