+++ src/MOI_wrapper.jl
@@ -11,6 +11,8 @@ export ∂, Δ, final, start, rem, Transition, add_transition!, guard, resetmaps
 
 @enum(TimeType, UNKNOWN, CONTINUOUS, DISCRETE, HYBRID)
 
+DEFAULT_MODE = 1
+
 mutable struct Optimizer <: MOI.AbstractOptimizer
     inner::Any
     nlp_model::Any
@@ -24,14 +26,14 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
     upper::Vector{Float64}
     start::Vector{MOI.Interval{Float64}}
     target::Vector{MOI.Interval{Float64}}
-    dynamic::Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}
+    dynamic::Dict{Int64, Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}} # mode_name -> [dynamic_func for each state] (in continuous and discrete time there is only one mode_name)
     obstacles::Vector{Tuple{Vector{MOI.VariableIndex}, MOI.HyperRectangle}}
     objective_sense::MOI.OptimizationSense
     objective_function::MOI.AbstractScalarFunction
     nonlinear_index::Int
-    modes_table::Vector{Union{Nothing, Tuple{JuMP.VariableRef, Int64, Function}}} # mode_index -> (mode_var, mode_val, mode_func)
-    guards_table::Vector{Union{Nothing, Tuple{JuMP.VariableRef, Int64, Int64, Function}}}  # guard_index -> (mode_var, src, dst, guard_func)
-    resetmaps_table::Vector{Union{Nothing, Tuple{JuMP.VariableRef, Int64, Int64, Function}}}  # resetmap_index -> (mode_var, src, dst, resetmap_func)
+    mode_variable::Union{Nothing, MOI.VariableIndex}
+    guards_table::Dict{String, Vector{Union{Nothing, Tuple{Int64, Int64, MOI.ScalarNonlinearFunction}}}}  # transition_index -> [(src, dst, guard_func) for each state]
+    resetmaps_table::Dict{String, Vector{Union{Nothing, Tuple{Int64, Int64, MOI.ScalarNonlinearFunction}}}}  # transition_index -> [(src, dst, resetmap_func) for each state]
     integer_variables::Dict{MOI.VariableIndex, Union{MOI.Integer, MOI.ZeroOne}}
     indicator_constraints::Vector{
         Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction},
@@ -50,14 +52,14 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
             Float64[],
             MOI.Interval{Float64}[],
             MOI.Interval{Float64}[],
-            Union{Nothing, MOI.ScalarNonlinearFunction}[],
+            Dict{Int64, Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}}(),
             Tuple{Vector{MOI.VariableIndex}, MOI.HyperRectangle}[],
             MOI.FEASIBILITY_SENSE,
             zero(MOI.ScalarAffineFunction{Float64}),
             0,
-            Union{Nothing, Tuple{JuMP.VariableRef, Int64, Function}}[],
-            Union{Nothing, Tuple{JuMP.VariableRef, Int64, Int64, Function}}[],
-            Union{Nothing, Tuple{JuMP.VariableRef, Int64, Int64, Function}}[],
+            nothing,
+            Dict{String, Vector{Union{Nothing, Tuple{Int64, Int64, MOI.ScalarNonlinearFunction}}}}(),
[…]
