### 2024-11-21 11d8434
CompatHelper: bump compat for LazySets to 3, (keep existing compat) (#396)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
