+++ Project.toml
@@ -54,7 +54,7 @@ IntervalArithmetic = "0.20, 0.21, 0.22"
 IntervalLinearAlgebra = "0.1"
 JuMP = "1"
 LaTeXStrings = "1"
-LazySets = "2"
+LazySets = "2, 3"
 Libtiff_jll = "4.4.0"
 LightXML = "0.9"
 MathOptInterface = "1"
