+++ src/MOI_wrapper.jl
@@ -13,91 +13,105 @@ export ∂, Δ, final, start, rem, TransitionAsModel, add_transition!, guard, re
 
 DEFAULT_MODE = 1
 
-mutable struct Optimizer <: MOI.AbstractOptimizer
-    inner::Any
+# Define the class Variable to handle a variable value, type, index, lower, upper, start, and target.
+mutable struct Variable
+    value::Float64
+    type::VariableType
+    index::Int
+    lower::Float64
+    upper::Float64
+    start::MOI.Interval{Float64}
+    target::MOI.Interval{Float64}
+
+    function Variable()
+        return new(0.0, INPUT, 0, -Inf, Inf, MOI.Interval(-Inf, Inf), MOI.Interval(-Inf, Inf))
+    end
+end
+
+# Define the class Objective to handle the sense and function.
+mutable struct Objective
+    sense::MOI.OptimizationSense
+    func::MOI.AbstractScalarFunction
+
+    function Objective()
+        return new(MOI.FEASIBILITY_SENSE, zero(MOI.ScalarAffineFunction{Float64}))
+    end
+end
+
+# Define the class Mode to handle nlp_model, evaluator, and dynamic.
+# The dynamic is a dictionary with mode_name -> [dynamic_func for each state]).
+# The evaluator is a dictionary with mode_name -> evaluator.
+# The nlp_model is a dictionary with mode_name -> nlp_model.
+# if no mode is specified, the default mode is used.
+mutable struct Mode
+    name::Int
     nlp_model::Any
     evaluator::Union{Nothing, MOI.Nonlinear.Evaluator}
+    dynamic::Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}
+
+    function Mode(name, nvars)
+        return new(name, MOI.Nonlinear.Model(), nothing, fill(nothing, nvars))
+    end
+end
+
+
+# Define the class Transition to handle source, destination, guard, and resetmap.
+struct Transition
+    source::Int
+    destination::Int
+    guard::Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}
+    resetmap::Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}
+
+    function Transition(src, dst, nvars)
+        return new(src, dst, fill(nothing, nvars), fill(nothing, nvars))
+    end
+end
+
+# Define the class Optimizer to handle the inner, modes, time_type, variables, derivative_values, obstacles, objective, nonlinear_index, mode_variable, transitions, integer_variables, and indicator_constraints.
+mutable struct Optimizer <: MOI.AbstractOptimizer
+    inner::Any
+    variables::Vector{Variable}
+    modes::Dict{Int, Mode}
     time_type::TimeType
-    variable_values::Vector{Float64}
     derivative_values::Vector{Float64}
-    variable_types::Vector{VariableType}
-    variable_index::Vector{Int} # MOI.VariableIndex -> state/input/mode index
-    lower::Vector{Float64}
-    upper::Vector{Float64}
-    start::Vector{MOI.Interval{Float64}}
-    target::Vector{MOI.Interval{Float64}}
-    dynamic::Dict{Int64, Vector{Union{Nothing, MOI.ScalarNonlinearFunction}}} # mode_name -> [dynamic_func for each state] (in continuous and discrete time there is only one mode_name)
     obstacles::Vector{Tuple{Vector{MOI.VariableIndex}, MOI.HyperRectangle}}
-    objective_sense::MOI.OptimizationSense
-    objective_function::MOI.AbstractScalarFunction
+    objective::Objective
     nonlinear_index::Int
     mode_variable::Union{Nothing, MOI.VariableIndex}
-    guards_table::Dict{
-        String,
[…]
