### 2024-11-29 c9f475e
implementing the partial hybrid system, uniformgrid is not supported, need to able to set algo
