+++ src/MOI_wrapper.jl
@@ -5,7 +5,8 @@ import JuMP
 import MathOptSymbolicAD
 import Symbolics
 
-export ∂, Δ, final, start, rem, TransitionAsModel, add_transition!, guard, resetmaps, themode
+export ∂,
+    Δ, final, start, rem, TransitionAsModel, add_transition!, guard, resetmaps, themode
 
 @enum(VariableType, INPUT, STATE, MODE)
 
@@ -24,7 +25,15 @@ mutable struct Variable
     target::MOI.Interval{Float64}
 
     function Variable()
-        return new(0.0, INPUT, 0, -Inf, Inf, MOI.Interval(-Inf, Inf), MOI.Interval(-Inf, Inf))
+        return new(
+            0.0,
+            INPUT,
+            0,
+            -Inf,
+            Inf,
+            MOI.Interval(-Inf, Inf),
+            MOI.Interval(-Inf, Inf),
+        )
     end
 end
 
@@ -54,7 +63,6 @@ mutable struct Mode
     end
 end
 
-
 # Define the class Transition to handle source, destination, guard, and resetmap.
 struct Transition
     source::Int
@@ -80,8 +88,10 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
     mode_variable::Union{Nothing, MOI.VariableIndex}
     transitions::Dict{String, Transition}
     integer_variables::Dict{MOI.VariableIndex, Union{MOI.Integer, MOI.ZeroOne}}
-    indicator_constraints::Vector{Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction}}
-    
+    indicator_constraints::Vector{
+        Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction},
+    }
+
     function Optimizer()
         return new(
             MOI.instantiate(Dionysos.Optim.Abstraction.UniformGridAbstraction.Optimizer),
@@ -95,7 +105,7 @@ mutable struct Optimizer <: MOI.AbstractOptimizer
             nothing,
             Dict{String, Transition}(),
             Dict{MOI.VariableIndex, Union{MOI.Integer, MOI.ZeroOne}}(),
-            Vector{Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction}}(),            
+            Vector{Tuple{MOI.VariableIndex, Bool, MOI.ScalarNonlinearFunction}}(),
         )
     end
 end
@@ -302,14 +312,13 @@ end
 function _update_guard_table!(model, src, dst, func, val)
     transition = _create_transition_if_not_exists!(model, src, dst)
 
-    model.transitions[transition].guard[val] = func
-
+    return model.transitions[transition].guard[val] = func
 end
 
 function _update_resetmaps_table!(model, src, dst, func, val)
     transition = _create_transition_if_not_exists!(model, src, dst)
 
-    model.transitions[transition].resetmap[val] = func
+    return model.transitions[transition].resetmap[val] = func
 end
 
 function MOI.add_constraint(
@@ -533,7 +542,13 @@ function TransitionAsModel(
     destination_mode::Int,
     switching_type::HybridSystems.AbstractSwitching = HybridSystems.ControlledSwitching(),
 )
-    return TransitionAsModel(model, mode_variable, source_mode, destination_mode, switching_type)
+    return TransitionAsModel(
+        model,
+        mode_variable,
+        source_mode,
+        destination_mode,
+        switching_type,
+    )
 end
 
 JuMP._valid_model(::TransitionAsModel, ::Any) = nothing
@@ -663,7 +678,11 @@ end
 
 function dynamic!(model, mode = DEFAULT_MODE)
[…]
