### 2024-12-03 982bb69
adding ConstrainedAffineControlContinuousSystem to bemporad_morari
