### 2024-12-03 2e8f965
adding hybrid support for the thermostat case
