+++ src/MOI_wrapper.jl
@@ -4,9 +4,21 @@ import MathematicalSystems, HybridSystems
 import JuMP
 import MathOptSymbolicAD
 import Symbolics
+import OSQP, HiGHS, Ipopt, Pavito, CDDLib
+using Polyhedra
+using FillArrays
 
 export ∂,
-    Δ, final, start, rem, TransitionAsModel, add_transition!, guard, resetmaps, themode
+    Δ,
+    final,
+    start,
+    rem,
+    TransitionAsModel,
+    add_transition!,
+    guard,
+    resetmaps,
+    themode,
+    set_algorithm!
 
 @enum(VariableType, INPUT, STATE, MODE)
 
@@ -723,7 +735,38 @@ function dynamic(model, x_idx, u_idx, mode = DEFAULT_MODE)
     return F_sys
 end
 
+function _rect(lb, ub, lib, T)
+    if lb == -Inf && ub == Inf
+        r =  FullSpace(1)
+    elseif lb == -Inf
+        r = hrep([HalfSpace([1], T(ub))])
+    elseif ub == Inf
+        r = hrep([HalfSpace([-1], -T(lb))])
+    else
+        r = HalfSpace([-1], -T(lb)) ∩ HalfSpace([1], T(ub))
+    end
+
+    return polyhedron(r, lib)
+end
+
+function _HyperRectangle_to_polyhedra(hrect, lib, T)
+    n = length(hrect.lb)#lower)  # number of dimensions
+    lb_vector = hrect.lb#lower
+    ub_vector = hrect.ub#upper
+
+    result = Vector{Polyhedra.Rep}(undef, n)
+    
+    for i in 1:n
+        result[i] = _rect(lb_vector[1], ub_vector[1], lib, T)
+    end
+    
+    return reduce(∩, result)
+end
+
 function _hybrid_system(model, x_idx, u_idx, _X_, _U_)
+    lib = CDDLib.Library()
+    T::Type = Float64
+
     println(">>Assembling the hybrid system")
 
     # Create the automaton
@@ -742,15 +785,32 @@ function _hybrid_system(model, x_idx, u_idx, _X_, _U_)
     # Assemble the continuous systems for each mode
     continuous_systems =
         Vector{MathematicalSystems.AbstractContinuousSystem}(undef, n_modes)
+
+    funcs = [
+        MathematicalSystems.ConstrainedAffineControlContinuousSystem(
+            zeros(Float64, length(x_idx), 1),
+            -0.5 .* ones(Float64, length(u_idx), 1),
+            fill(0.0, 1),
+            _X_,
+            _U_,
+        ),
+        MathematicalSystems.ConstrainedAffineControlContinuousSystem(
+            zeros(Float64, length(x_idx), 1),
+            zeros(Float64, length(u_idx), 1),
+            fill(1.0, 1),
+            _X_,
+            _U_,
+        )
+    ]
     for mode in eachindex(model.modes)
-        continuous_systems[mode] =
-            MathematicalSystems.ConstrainedBlackBoxControlContinuousSystem(
+        continuous_systems[mode] = funcs[mode]
+            #=MathematicalSystems.ConstrainedBlackBoxControlContinuousSystem(
                 dynamic(model, x_idx, u_idx, mode),
                 Dionysos.Utils.get_dims(_X_),
                 Dionysos.Utils.get_dims(_U_),
                 _X_,
                 _U_,
-            )
+            )=#
     end
 
     # Assemble the guard for each transition
@@ -758,19 +818,27 @@ function _hybrid_system(model, x_idx, u_idx, _X_, _U_)
     guard_constraints = [
         Dionysos.Utils.HyperRectangle([-Inf], [19.0]),
         Dionysos.Utils.HyperRectangle([21.0], [Inf]),
[…]
