+++ src/optim/abstraction/uniform_grid_abstraction.jl
@@ -7,12 +7,6 @@ import StaticArrays
 import MathematicalSystems
 
 import Dionysos
-const DI = Dionysos
-const UT = DI.Utils
-const DO = DI.Domain
-const ST = DI.System
-const SY = DI.Symbolic
-const PR = DI.Problem
 
 @enum ApproxMode GROWTH LINEARIZED DELTA_GAS
 
@@ -24,14 +18,14 @@ using JuMP
 Solver based on the classical abstraction method (used for instance in SCOTS) for which the whole domain is partioned into hyper-rectangular cells, independently of the control task.
 """
 mutable struct Optimizer{T} <: MOI.AbstractOptimizer
-    concrete_problem::Union{Nothing, PR.ProblemType}
-    abstract_problem::Union{Nothing, PR.OptimalControlProblem, PR.SafetyProblem}
-    abstract_system::Union{Nothing, SY.SymbolicModelList}
-    abstract_controller::Union{Nothing, UT.SortedTupleSet{2, NTuple{2, Int}}}
+    concrete_problem::Union{Nothing, Dionysos.Problem.ProblemType}
+    abstract_problem::Union{Nothing, Dionysos.Problem.OptimalControlProblem, Dionysos.Problem.SafetyProblem}
+    abstract_system::Union{Nothing, Dionysos.Symbolic.SymbolicModelList}
+    abstract_controller::Union{Nothing, Dionysos.Utils.SortedTupleSet{2, NTuple{2, Int}}}
     concrete_controller::Any
     discretized_system::Any
-    state_grid::Union{Nothing, DO.Grid}
-    input_grid::Union{Nothing, DO.Grid}
+    state_grid::Union{Nothing, Dionysos.Domain.Grid}
+    input_grid::Union{Nothing, Dionysos.Domain.Grid}
     jacobian_bound::Union{Nothing, Function}
     time_step::T
     num_sub_steps_system_map::Int
@@ -154,11 +148,11 @@ function build_abstraction(concrete_system, model)
         error("Please set the `input_grid`.")
     end
 
-    Xfull = DO.DomainList(model.state_grid)
-    DO.add_set!(Xfull, concrete_system.X, DO.INNER)
-    Ufull = DO.DomainList(model.input_grid)
-    DO.add_set!(Ufull, concrete_system.U, DO.CENTER)
-    abstract_system = SY.NewSymbolicModelListList(Xfull, Ufull)
+    Xfull = Dionysos.Domain.DomainList(model.state_grid)
+    Dionysos.Domain.add_set!(Xfull, concrete_system.X, Dionysos.Domain.INNER)
+    Ufull = Dionysos.Domain.DomainList(model.input_grid)
+    Dionysos.Domain.add_set!(Ufull, concrete_system.U, Dionysos.Domain.CENTER)
+    abstract_system = Dionysos.Symbolic.NewSymbolicModelListList(Xfull, Ufull)
 
     # TODO add noise to the description of the system so in a MathematicalSystems
     #      this is a workaround
@@ -168,12 +162,12 @@ function build_abstraction(concrete_system, model)
     model.discretized_system = discretized_system(concrete_system, model, noise)
 
     if model.δGAS
-        @time SY.compute_deterministic_symmodel_from_controlsystem!(
+        @time Dionysos.Symbolic.compute_deterministic_symmodel_from_controlsystem!(
             abstract_system,
             model.discretized_system,
         )
     else
-        @time SY.compute_symmodel_from_controlsystem!(
+        @time Dionysos.Symbolic.compute_symmodel_from_controlsystem!(
             abstract_system,
             model.discretized_system,
         )
@@ -182,18 +176,18 @@ function build_abstraction(concrete_system, model)
 end
 
 function build_abstract_problem(
[…]
