+++ src/optim/abstraction/uniform_grid_abstraction.jl
@@ -75,17 +75,76 @@ function MOI.get(model::Optimizer, param::MOI.RawOptimizerAttribute)
     return getproperty(model, Symbol(param.name))
 end
 
-function discretized_system(
+"""
+    validate_model!(model::Optimizer, required_fields::Vector{Symbol})
+
+Ensures the specified `required_fields` are set in the `model`. Throws an error
+if any field is missing.
+"""
+function _validate_model(model::Optimizer, required_fields::Vector{Symbol})
+    for field in required_fields
+        if isnothing(getfield(model, field))
+            error("Please set the `$(field)`.")
+        end
+    end
+end
+
+"""
+    validate_continuous_model!(model::Optimizer)
+
+Validates that the `Optimizer` model is correctly configured for continuous
+systems. Throws an error if required fields like `time_step` or `jacobian_bound`
+are missing or invalid.
+"""
+function _validate_continuous_model(model::Optimizer)
+    _validate_model(model, [:time_step, :jacobian_bound])
+    if isnan(model.time_step)
+        error("Please set the `time_step`.")
+    end
+end
+
+"""
+    validate_discrete_model!(model::Optimizer)
+
+Validates that the `Optimizer` model is correctly configured for discrete
+systems. Throws an error if required fields like `sys_inv_map` or `growthbound_map`
+are missing or invalid.
+"""
+function _validate_discrete_model(model::Optimizer)
+    _validate_model(model, [:growthbound_map, :sys_inv_map])
+end
+
+"""
+    discretized_system(concrete_system, model, noise)
+
+Returns the discretized system based on the `concrete_system` and `model`.
+"""
+function _maybe_discretized_system(
+    concrete_system,
+    model,
+    noise,
+)
+    if isa(concrete_system, MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem)
+        _validate_discrete_model(model)
+        return _discrete_system(concrete_system, model, noise)
+    elseif isa(concrete_system, MathematicalSystems.ConstrainedBlackBoxControlContinuousSystem)
+        _validate_continuous_model(model)
+        return _discretize_continuous_system(concrete_system, model, noise)
+    else
+        error("Unsupported system type: $(typeof(concrete_system))")
+    end
+end
+
+"""
+    discretized_system(concrete_system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem, model, noise)
+
+Returns the discretized system based on the `concrete_system` and `model`.
+"""
+function _discrete_system(
     concrete_system::MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem,
     model,
     noise,
 )
-    if isnothing(model.growthbound_map)
-        error("Please set the `growthbound_map`.")
-    end
-    if isnothing(model.sys_inv_map)
-        error("Please set the `sys_inv_map`.")
-    end
     return Dionysos.System.ControlSystemGrowth(
         1.0, # `time_step` should be ignored by `concrete_system.f`, `model.growthbound_map` and `model.sys_inv_map` anyway
         noise,
@@ -96,19 +155,17 @@ function discretized_system(
     )
 end
 
-function discretized_system(
+"""
+    _discretize_continuous_system(concrete_system::MathematicalSystems.ConstrainedBlackBoxControlContinuousSystem, model, noise)
+
+Returns the discretized system based on the `concrete_system` and `model`.
+"""
+function _discretize_continuous_system(
     concrete_system::MathematicalSystems.ConstrainedBlackBoxControlContinuousSystem,
     model,
     noise,
 )
-    if isnan(model.time_step)
[…]
