### 2024-12-03 378749b
refactoring abstraction building
