### 2024-12-10 #160 @lucaferranti
@adrienbanse  to finalize the acceptance, please see the "additional author tasks after review is complete" in the message above ☝️

The critical thing is to have an up-to-date release of the software and a corresponding archive in zenodo. Once this is done, please post here the version and DOI here.

Also, looking at the paper I see some parts are highlighted in red (diffs for reviewer I presume). Now that review is concluded, could you harmonize the text?
