### 2024-12-11 #160 @adrienbanse
@schillic @jmskov Thanks a lot for the feedback!
@lucaferranti Many thanks for your work :pray:. We will implement @schillic last comments very soon, and complete the checklist you just generated.
