+++ paper/paper.tex
@@ -76,7 +76,7 @@ This paper is structured as follows.  Section~\ref{sec:abstraction} introduces t
 % Intervals
 The sets $\R,\Z, \Z_+$ denote respectively the sets of real numbers, integers and non-negative integers.
 % Kleene closure
-\color{red}Given a set $A$, the set $A^*$ is its \emph{Kleene closure}. \color{black}
+Given a set $A$, the set $A^*$ is its \emph{Kleene closure}. 
 % Set-valued map
 Given two sets $A,B$, we define a \emph{single-valued map} as $f:A\rightarrow B$, while a~\emph{set-valued map} is defined as $f:A\rightarrow 2^B$, where $2^B$ is the power set of $B$, i.e., the set of all subsets of $B$. 
 The image of a subset $\Omega\subseteq A$ under $f:A\rightarrow 2^B$ is denoted $f(\Omega)$.
@@ -122,7 +122,7 @@ A \emph{finite-state} system, in contrast to an \emph{infinite-state} system, re
 
 \vskip 6pt
 % Definition of a solution
-A tuple $(\seq x, \seq u)\in \set{X}^{T+1}\times \set{U}^{T}$ is a \emph{trajectory} of length $T$ of the system $\Sys=(\set{X},\set{U}, F)$ starting at $x(0)$ if $T\in\N\cup \{\infty\}$, $x(0)\in\set{X}$, $\forall k\in \{0,\ldots,\textcolor{red}{T-1}\}: u(k)\in \set{U}(x(k))$ and $x(k+1) \in F(x(k), u(k))$.
+A tuple $(\seq x, \seq u)\in \set{X}^{T+1}\times \set{U}^{T}$ is a \emph{trajectory} of length $T$ of the system $\Sys=(\set{X},\set{U}, F)$ starting at $x(0)$ if $T\in\N\cup \{\infty\}$, $x(0)\in\set{X}$, $\forall k\in \{0,\ldots,T-1\}: u(k)\in \set{U}(x(k))$ and $x(k+1) \in F(x(k), u(k))$.
 The set of trajectories of $\Sys$ is called the \emph{behavior} of $\Sys$, denoted~$\set{B}(\Sys)$.
 
 \vskip 6pt 
@@ -139,7 +139,7 @@ We now define \emph{static} controllers, which are characterized by the fact tha
 %\vskip 6pt
 We now define the control problem.
 \begin{defi}\label{def:specification}
-Consider a system $\Sys = (\set{X},\set{U}, F)$. A \emph{specification} $\Sigma$ for $\Sys$ is defined as any subset \color{red}$\Sigma \subseteq (\set{X}\times \set{U})^{*} \cup (\set{X}\times \set{U})^{\infty}$\color{black}. It is said that system $\Sys$ \emph{satisfies} the specification $\Sigma$ if $\set{B}(\Sys)\subseteq \Sigma$. 
+Consider a system $\Sys = (\set{X},\set{U}, F)$. A \emph{specification} $\Sigma$ for $\Sys$ is defined as any subset $\Sigma \subseteq (\set{X}\times \set{U})^{*} \cup (\set{X}\times \set{U})^{\infty}$. It is said that system $\Sys$ \emph{satisfies} the specification $\Sigma$ if $\set{B}(\Sys)\subseteq \Sigma$. 
 A system $\Sys$ together with a specification $\Sigma$ constitute a \emph{control problem} $(\Sys,\Sigma)$.
 Additionally, a controller $\Cont$ is said to \emph{solve} the control problem $(\Sys,\Sigma)$ if $\Cont\times \Sys$ satisfies the specification $\Sigma$.
 \end{defi}
[…]
