+++ paper/header.tex
@@ -4,10 +4,9 @@
 
 \author[1]{Julien Calbert}
 \author[1]{Adrien Banse}
-\author[2]{Benoît Legat}
+\author[1]{Benoît Legat}
 \author[1]{Raphaël M. Jungers}
 \affil[1]{ICTEAM, UCLouvain}
-\affil[2]{STADIUS, KU Leuven}
 
 \keywords{Systems and Control, Symbolic Models, Smart abstractions}
 
