### 2024-12-12 #160 @adrienbanse
Dear @lucaferranti, we have published the release on Zenodo, here is the DOI https://doi.org/10.5281/zenodo.14421121, and made sure that everything goes according to the checklist you provided.
We have also implemented the latest comments of @schillic, and removed the red text.

Everything should be ready on our side!

cc @blegat @JulienCalbert @raphaeljungers
