### 2024-12-12 #160 @lucaferranti
@adrienbanse just double checking the version associated with the release and paper is 0.1.3, can you confirm?
