### 2024-12-12 #160 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
