### 2024-12-12 #160 @lucaferranti
@adrienbanse this is ready :tada: if you could have  a final look at the paper and confirm it looks good, I can then pull the trigger
