### 2024-12-13 #160 @adrienbanse
@lucaferranti Just had a last check, it seems good to me! Just a tiny question: is it normal that the ORCID do not appear on the paper?
