### 2024-12-13 #160 @lucaferranti
On the paper pdf? Yeah, that's how the template works atm. Authors orcid are included in the paper webpage when published, see for example https://proceedings.juliacon.org/papers/10.21105/jcon.00164
